% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASNR_Fig08x1a.R
\docType{data}
\name{ASNR_Fig08x1a}
\alias{ASNR_Fig08x1a}
\title{ASNR_Fig08x1a}
\format{
A network object of class "matrix".
}
\usage{
ASNR_Fig08x1a
}
\description{
This is an example dataset used in Borgatti, Everett, Johnson and Agneessens (2022).
}
\references{
{Borgatti, Everett, Johnson and Agneessens (2022) Analyzing social networks with R. Sage.}
}
\keyword{datasets}
