% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASNR_Fig09x2.R
\docType{data}
\name{ASNR_Fig09x2}
\alias{ASNR_Fig09x2}
\title{ASNR_Fig09x2}
\format{
A network object of class "matrix".
}
\usage{
ASNR_Fig09x2
}
\description{
This is an example dataset used in Borgatti, Everett, Johnson and Agneessens (2022).
}
\references{
{Borgatti, Everett, Johnson and Agneessens (2022) Analyzing social networks with R. Sage.}
}
\keyword{datasets}
