% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASNR_Fig10x3B.R
\docType{data}
\name{ASNR_Fig10x3B}
\alias{ASNR_Fig10x3B}
\title{ASNR_Fig10x3B}
\format{
A network object of class "matrix".
}
\usage{
ASNR_Fig10x3B
}
\description{
This is an example dataset used in Borgatti, Everett, Johnson and Agneessens (2022).
}
\references{
{Borgatti, Everett, Johnson and Agneessens (2022) Analyzing social networks with R. Sage.}
}
\keyword{datasets}
