% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baker_journals}
\alias{baker_journals}
\title{Baker's social work journals}
\format{
An \code{xucinet} object, 20 x 20, one valued directed relation.
}
\source{
Baker, D. R. (1992). A structural analysis of social work journal
network: 1985-1986. \emph{Journal of Social Service Research}, 15(3-4), 153-168.
}
\usage{
baker_journals
}
\description{
Citations among 20 social work journals for 1985-1986: cell (i, j) is the
number of citations from articles in journal i to articles in journal j.
Baker found a core-periphery structure, and argued that American journals
are less likely to engage the international literature.
}
\section{Book sections}{
 12.8 (practice 12.5).
}

\examples{
xdensity(baker_journals)
}
\keyword{datasets}
