% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernard_HamRadio.R
\docType{data}
\name{Bernard_HamRadio}
\alias{Bernard_HamRadio}
\title{Amateur HAM radio calls by Bernard and colleagues.}
\format{
A list object (xUCINET project/dataset) containing 2 network and no attributes
\describe{
\item{$ProjectInfo}{-  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{-  A data.frame with attributes for nodes}
\item{$NetworkInfo}{-  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{-  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{-  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Bernard_HamRadio
}
\description{
Bernard & Killworth, later with the help of Sailer, collected five sets of data on human interactions in bounded groups and on the actors' ability to recall those interactions. In each study they obtained measures of social interaction among all actors, and ranking data based on the subjects' memory of those interactions. This dataset is based on amateur HAM radio calls made over a one-month period.
}
\examples{
Bernard_HamRadio

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
