% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borgatti_Camp92.R
\docType{data}
\name{Borgatti_Camp92}
\alias{Borgatti_Camp92}
\title{Network among participants of a workshop}
\format{
A list object (xUCINET project/dataset) containing 2 network and 2 attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Borgatti_Camp92
}
\description{
These data were collected by Steve Borgatti, Russ Bernard, Bert Pelto and Gery Ryan at the 1992 NSF Summer Institute on Research Methods in Cultural Anthropology. This was a 3-week course given to 14 carefully selected participants. Network data were collected at the end of each week. These data were collected at the end of the second week. The data were collected by placing each person's name on a card and asking each respondent to sort the cards in order of how much interaction they had with that person since the beginning of the course (known informally as 'camp'). This results in rank order data in which a '1' indicates the most interaction while a '17' indicates the least interaction.
}
\examples{
Borgatti_Camp92

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
