% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borgatti_Scientists504.R
\docType{data}
\name{Borgatti_Scientists504}
\alias{Borgatti_Scientists504}
\title{Collaboration among 504 scientists.}
\format{
A list object (xUCINET project/dataset) containing 1 networks and 3 attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Borgatti_Scientists504
}
\description{
This is the main component of the Borgatti_Scientists960 dataset.
}
\examples{
Borgatti_Scientists504

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
