% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borgatti_Scientists960.R
\docType{data}
\name{Borgatti_Scientists960}
\alias{Borgatti_Scientists960}
\title{Collaboration among 960 scientists.}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Borgatti_Scientists960
}
\description{
These data indicate the amount of time (in days, and rounded to the nearest day) that scientists at a large applied research institute worked on projects together. The data are derived from a 2-mode matrix (960 persons by 3154 projects) in which the cells indicated how much time each scientist worked on each project. The 1-mode projection was achieved by the sum of minimums method. For example, if person A spent 1 day on project 1 and 3 days on project 2, and person B spent 5 days on project 1 and 2 days on project 2, the amount of time worked together is 1+2 = 3 days.
See $ProjectInfo for more information.
}
\examples{
Borgatti_Scientists960

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
