% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burkhardt_GovernmentAgency.R
\docType{data}
\name{Burkhardt_GovernmentAgency}
\alias{Burkhardt_GovernmentAgency}
\title{Advice among employees in a government agency collected by Burkhardt and Brass.}
\format{
A list object (xUCINET project/dataset) containing 5 networks and no attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Burkhardt_GovernmentAgency
}
\description{
Network data capturing who communicated with whom about work matters at a government agency at five points in tie. The data were collected by Burkhardt and Brass on 75 employees.
}
\examples{
Burkhardt_GovernmentAgency

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
