% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camp92_attr}
\alias{camp92_attr}
\title{Camp 92 participant attributes}
\format{
A data frame, 18 rows keyed by participant name, 3 columns:
\describe{
\item{Gender}{1 = female, 2 = male}
\item{Role}{1 = participant, 2 = instructor}
\item{Betweenness}{betweenness centrality, as reported in the book}
}
}
\source{
As \link{camp92}.
}
\usage{
camp92_attr
}
\description{
Camp 92 participant attributes
}
\examples{
head(camp92_attr)
}
\seealso{
\link{camp92}
}
\keyword{datasets}
