% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{campnet}
\alias{campnet}
\title{Campnet}
\format{
An \code{xucinet} object, 18 x 18, one binary directed relation.
}
\source{
Borgatti, S. P., Bernard, H. R., Pelto, P., Ryan, G. and DeJordy, R.
(2012). The Camp '92 dataset. \emph{Connections}, 32(1), 51-53.
}
\usage{
campnet
}
\description{
The binary interaction network among the same 18 people as \link{camp92}, used as
the running example through much of the book. A tie means the respondent
placed that person among those they interacted with most.
}
\section{Book sections}{
 5.2.3.
}

\examples{
xdensity(campnet)
}
\seealso{
\link{camp92}, \link{camp92_attr}
}
\keyword{datasets}
