% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{davis}
\alias{davis}
\title{Davis's southern women}
\format{
An \code{xucinet} object, 18 women x 14 events, 2-mode, binary.
}
\source{
Davis, A., Gardner, B. and Gardner, M. (1941). \emph{Deep South}. Chicago:
University of Chicago Press. See also Breiger, R. (1974). The duality of
persons and groups. \emph{Social Forces}, 53, 181-190.
}
\usage{
davis
}
\description{
The canonical two-mode dataset: attendance by 18 women at 14 social events in
a Mississippi town in the 1930s. Cell (i, j) is 1 if woman i attended event
j. Breiger used it to make the argument about the duality of persons and
groups, and it has been a benchmark for two-mode methods ever since.
}
\section{Book sections}{
 2.5, 3.5, 13.1-13.7.2 (practices 13.1-13.5).
}

\examples{
xdensity(davis)
}
\keyword{datasets}
