% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eies_attr}
\alias{eies_attr}
\title{EIES researcher attributes}
\format{
A data frame, 32 rows keyed by researcher id, 2 columns:
\describe{
\item{Citations}{citations to the person's work in the Social Science
Citation Index at the start of the study}
\item{Discipline}{1 = sociology, 2 = anthropology,
3 = mathematics/statistics, 4 = other}
}
}
\source{
As \link{eies}.
}
\usage{
eies_attr
}
\description{
EIES researcher attributes
}
\examples{
table(eies_attr$Discipline)
}
\seealso{
\link{eies}
}
\keyword{datasets}
