% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Greenacre_DoctoratesByYearAndField.R
\docType{data}
\name{Greenacre_DoctoratesByYearAndField}
\alias{Greenacre_DoctoratesByYearAndField}
\title{Matrix containing number of doctorates by field and year.}
\format{
A rectangular matrix
}
\source{
{UCINET}
}
\usage{
Greenacre_DoctoratesByYearAndField
}
\description{
Matrix containing number of doctorates by field and year.
}
\examples{
Greenacre_DoctoratesByYearAndField
}
\keyword{datasets}
