% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hawthorne_BankWiring.R
\docType{data}
\name{Hawthorne_BankWiring}
\alias{Hawthorne_BankWiring}
\title{Interactions among employees in a bank wiring room (Hawthorne plant).}
\format{
A list object (xUCINET project/dataset) containing 6 networks and no attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Hawthorne_BankWiring
}
\description{
These are the observational data on 14 Western Electric (Hawthorne Plant) employees from the bank wiring room first presented in Roethlisberger & Dickson (1939). The data are better known through a scrutiny made of the interactions in Homans (1950), as well as the CONCOR analysis presented in Breiger et al (1975). The employees worked in a single room and include two inspectors (I1 and I3), three solderers (S1, S2 and S3), and nine wiremen or assemblers (W1 to W9). Contains six types of interactions among employees.
}
\examples{
Hawthorne_BankWiring

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
