% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hightech_attr}
\alias{hightech_attr}
\title{High-tech manager attributes}
\format{
A data frame, 21 rows keyed by manager id, 4 columns:
\describe{
\item{Age}{age in years}
\item{Tenure}{length of service in years}
\item{Level}{1 = CEO, 2 = vice president, 3 = manager}
\item{Department}{department code 1-4, with the CEO in department 0}
}
}
\source{
As \link{hightech}.
}
\usage{
hightech_attr
}
\description{
High-tech manager attributes
}
\examples{
table(hightech_attr$Level)
}
\seealso{
\link{hightech}
}
\keyword{datasets}
