% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kapferer_TailorShop.R
\docType{data}
\name{Kapferer_TailorShop}
\alias{Kapferer_TailorShop}
\title{Interactions among workers in a tailor shop school kids in a school collected by Kapferer.}
\format{
An object of class \code{list} of length 7.
}
\source{
{UCINET}
}
\usage{
Kapferer_TailorShop
}
\description{
Bruce Kapferer (1972) observed interactions in a tailor shop in Zambia (then Northern Rhodesia) over a period of ten months. His focus was the changing patterns of alliance among workers during extended negotiations for higher wages. The matrices represent two different types of interaction, recorded at two different times (seven months apart) over a period of one month: 'instrumental' (work- and assistance-related) interactions at the two times; and 'sociational' (friendship, socioemotional) interactions. The data are particularly interesting since an abortive strike occurred after the first set of observations, and a successful strike took place after the second.
See $ProjectInfo for more information.
}
\examples{
Kapferer_TailorShop

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
