% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knecht_attr}
\alias{knecht_attr}
\title{Knecht pupil attributes}
\format{
A data frame, 26 rows keyed by pupil id, 14 columns: \code{Sex2M}
(2 = male), \code{Age} at wave 1, \code{Ethnicity}, \code{Religion}, \code{DelinquencyT1} to
\code{DelinquencyT4}, \code{AlcoholT2} to \code{AlcoholT4} (alcohol was not measured at
wave 1), \code{CapacityT0}, \code{PresenceStarts} and \code{PresenceEnds}.
}
\source{
As \link{knecht}.
}
\usage{
knecht_attr
}
\description{
Knecht pupil attributes
}
\examples{
head(knecht_attr)
}
\seealso{
\link{knecht}
}
\keyword{datasets}
