% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lazega}
\alias{lazega}
\title{Lazega's corporate law firm}
\format{
An \code{xucinet} object, 71 x 71, three relations:
\describe{
\item{Advice}{colleagues gone to for non-technical professional advice}
\item{Coworking}{colleagues worked with on at least one case}
\item{Friendship}{colleagues socialised with outside work}
}
}
\source{
Lazega, E. (2001). \emph{The collegial phenomenon: the social mechanisms
of cooperation among peers in a corporate law partnership}. Oxford
University Press. See also Snijders, T. A. B., Pattison, P. E., Robins,
G. L. and Handcock, M. S. (2006). New specifications for exponential random
graph models. \emph{Sociological Methodology}, 99-153.
}
\usage{
lazega
}
\description{
Advice, co-working and friendship among 71 attorneys, partners and associates
alike, in a north-eastern US corporate law firm between 1988 and 1991. A
standard testbed for exponential random graph models, because the three
relations and the rich attribute set let you ask what predicts a tie.
}
\examples{
xrelations(lazega)
}
\seealso{
\link{lazega_attr}
}
\keyword{datasets}
