% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lazega_attr}
\alias{lazega_attr}
\title{Law firm attorney attributes}
\format{
A data frame, 71 rows keyed by attorney id, 7 columns:
\describe{
\item{Status}{1 = partner, 2 = associate}
\item{Gender}{1 = male, 2 = female}
\item{Office}{1 = Boston, 2 = Hartford, 3 = Providence}
\item{Tenure}{years with the firm}
\item{Age}{age in years}
\item{Practice}{1 = litigation, 2 = corporate}
\item{LawSchool}{law school attended, as a code}
}
}
\source{
As \link{lazega}.
}
\usage{
lazega_attr
}
\description{
Law firm attorney attributes
}
\examples{
table(lazega_attr$Office)
}
\seealso{
\link{lazega}
}
\keyword{datasets}
