% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainas_terro}
\alias{mainas_terro}
\title{Mainas terrorist contact network}
\format{
An \code{xucinet} object, 4275 x 4275, one valued relation.
}
\source{
Mainas, E. D. (2012). The analysis of criminal and terrorist
organisations as social network structures: a quasi-experimental study.
\emph{International Journal of Police Science and Management}, 14(3), 264-282.
}
\usage{
mainas_terro
}
\description{
Known contacts among 4275 people, obtained from Europol by Mainas and
analysed by its Counter-Terrorism and Analysis units. The kind of tie was not
disclosed. The largest network the package ships, and a realistic test of
routines on a sparse graph.
}
\section{Book sections}{
 11.7 (ASNR).
}

\examples{
dim(mainas_terro)
}
\keyword{datasets}
