% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAT_12_5.R
\docType{data}
\name{MAT_12_5}
\alias{MAT_12_5}
\title{Mat_12_5}
\format{
A network object of class "matrix".
}
\usage{
MAT_12_5
}
\description{
This is an example dataset used in Borgatti, Everett, Johnson and Agneessens (2022).
}
\references{
{Borgatti, Everett, Johnson and Agneessens (2022) Analyzing social networks with R. Sage.}
}
\keyword{datasets}
