% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newcomb_Fraternity.R
\docType{data}
\name{Newcomb_Fraternity}
\alias{Newcomb_Fraternity}
\title{Network ties collected by Newcomb among a fraternity.}
\format{
An object of class \code{list} of length 18.
}
\source{
{UCINET}
}
\usage{
Newcomb_Fraternity
}
\description{
These 15 matrices record weekly sociometric preference rankings from 17 men attending the University of Michigan in the fall of 1956;  data from week 9 are missing. A '1' indicates first preference, and no ties were allowed. The men were recruited to live in off-campus (fraternity) housing, rented for them as part of the Michigan Group Study Project supervised by Theodore Newcomb from 1953 to 1956. All were incoming transfer students with no prior acquaintance of one another.
See $ProjectInfo for more information.
}
\examples{
Newcomb_Fraternity

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
