% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newguinea}
\alias{newguinea}
\title{Read's New Guinea highland tribes}
\format{
An \code{xucinet} object, 16 x 16, two relations \code{Alliance} and
\code{Opposition}, both binary and symmetric.
}
\source{
Read, K. (1954). Cultures of the central highlands, New Guinea.
\emph{Southwestern Journal of Anthropology}, 10, 1-43. See also Everett, M. G.
and Borgatti, S. P. (2014). Networks containing negative ties. \emph{Social
Networks}, 38, 111-120.
}
\usage{
newguinea
}
\description{
Alliance and opposition among 16 Gahuku-Gama sub-tribes in the central
highlands of New Guinea. Because it records positive and negative ties on the
same set of actors, it is the standard example for methods that handle signed
networks.
}
\examples{
xrelations(newguinea)
}
\keyword{datasets}
