% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{padgett_attr}
\alias{padgett_attr}
\title{Florentine family attributes}
\format{
A data frame, 16 rows keyed by family name, 3 columns:
\describe{
\item{Wealth}{net wealth in 1427, in thousands of lira}
\item{NoPriors}{seats on the civic council held between 1282 and 1344}
\item{NoTies}{total business or marriage ties in the full 116-family data}
}
}
\source{
As \link{padgett}.
}
\usage{
padgett_attr
}
\description{
Florentine family attributes
}
\examples{
padgett_attr[order(-padgett_attr$Wealth), ][1:5, ]
}
\seealso{
\link{padgett}
}
\keyword{datasets}
