% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pane_training}
\alias{pane_training}
\title{Pane's training squad}
\format{
An \code{xucinet} object, 23 x 23, one binary directed relation.
}
\source{
Pane, A. (2006). \emph{Examining newcomers' socialization into an
organization}. Unpublished DPsych thesis, University of Melbourne. See also
Robins, G., Pattison, P. and Wang, P. (2009). \emph{Social Networks}, 31(2),
105-117.
}
\usage{
pane_training
}
\description{
Who trusts whom among 23 members of a training squad, collected in a study of
how newcomers are socialised into an organisation.
}
\section{Book sections}{
 15.4 (ASNR).
}

\examples{
xdensity(pane_training)
}
\keyword{datasets}
