% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{papuan_village}
\alias{papuan_village}
\title{Papuan village taro exchange}
\format{
An \code{xucinet} object, 22 x 22, one binary symmetric relation.
}
\source{
Schwimmer, E. (1973). \emph{Exchange in the social structure of the
Orokaiva}. New York: St Martin's. See also Hage, P. and Harary, F. (1983).
\emph{Structural models in anthropology}. Cambridge University Press.
}
\usage{
papuan_village
}
\description{
Gift-giving, specifically the exchange of taro, among 22 households in an
Orokaiva village in Papua. Schwimmer's point is that these ties define who
may properly mediate a request for help. Called Taro in UCINET.
}
\section{Book sections}{
 12.6 (ASNR).
}

\examples{
xdensity(papuan_village)
}
\keyword{datasets}
