% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schwimmer_PapuanVillage.R
\docType{data}
\name{Schwimmer_PapuanVillage}
\alias{Schwimmer_PapuanVillage}
\title{Schwimmer's gift-giving network}
\format{
A list object (xUCINET project/dataset) containing 1 network and no attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Schwimmer_PapuanVillage
}
\description{
These data represent the relation of gift-giving (taro exchange) among 22 households in a Papuan village. Hage & Harary (1983) used them to illustrate a graph Hamiltonian cycle. Schwimmer points out how these ties function to define the appropriate persons to mediate the act of asking for or receiving assistance among group members. This dataset is called 'Taro' in Windows Ucinet.
}
\examples{
Schwimmer_PapuanVillage

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
