% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade_pre29}
\alias{trade_pre29}
\title{Trade flows before the 1929 crash}
\format{
An \code{xucinet} object, 15 x 15, one valued directed relation, in
percentage of exports to each other country.
}
\source{
Savage, I. R. and Deutsch, K. W. (1960). A statistical model of the
gross analysis of transaction flows. \emph{Econometrica}, 28, 551-572.
}
\usage{
trade_pre29
}
\description{
Transaction flows between 15 countries in the years before the 1929 crash,
as used by Savage and Deutsch to argue for a statistical baseline against
which observed trade should be judged.
}
\section{Book sections}{
 7.2.1, 7.2.2 (ASNR).
}

\examples{
rownames(as.matrix(trade_pre29))
}
\keyword{datasets}
