% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolfe_primates}
\alias{wolfe_primates}
\title{Wolfe's primates}
\format{
An \code{xucinet} object, 20 x 20, two relations:
\describe{
\item{Kinship}{putative kin relationships among the animals}
\item{JointPresence}{joint presence at the river, summed over all pairs}
}
}
\source{
Borgatti, S. P., Everett, M. G., Johnson, J. C. and Agneessens, F.
\emph{Analyzing Social Networks}. 3rd edition. Sage.
}
\usage{
wolfe_primates
}
\description{
Three months of interaction among a troop of monkeys observed in the wild by
Linda Wolfe as they sported by a river in Ocala, Florida.
}
\section{Book sections}{
 7.2.1, 7.2.2 (ASNR).
}

\examples{
xrelations(wolfe_primates)
}
\seealso{
\link{wolfe_primates_attr}
}
\keyword{datasets}
