% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolfe_primates_attr}
\alias{wolfe_primates_attr}
\title{Primate attributes}
\format{
A data frame, 20 rows keyed by animal id, 3 columns:
\describe{
\item{Age}{age in years}
\item{Sex}{1 = male, 2 = female}
\item{Rank}{rank in the troop, 1 being highest}
}
}
\source{
As \link{wolfe_primates}.
}
\usage{
wolfe_primates_attr
}
\description{
Primate attributes
}
\examples{
head(wolfe_primates_attr)
}
\seealso{
\link{wolfe_primates}
}
\keyword{datasets}
