% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{xAddAttributesToProject}
\alias{xAddAttributesToProject}
\title{Adds nodal attributes to an existing "dataproject" object}
\usage{
xAddAttributesToProject(
  ProjectName,
  ATTFILE1 = NULL,
  FileType = NULL,
  AttributesDescription = NULL,
  Mode = NA
)
}
\arguments{
\item{ProjectName}{The name of the "dataproject", to which the attribute needs to be added.}

\item{ATTFILE1}{An attribute file that needs to be added. This could be a csv file or an R object (data.frame). The input type is defined in "FileType". If reading in a csv file, the first column should contain the node names (labels), so the function can check that the names correspond with the existing data. If it is an R object, these names should be in the rownames.}

\item{FileType}{The type of data that is being imported. Default is "csv". Other options are "Robject" when the data is already an object in R, and "csv2". Note that by default a csv file use a point (".") for the decimal point and a comma for the separator in most countries. However, in some Western European countries a comma is used for the decimal point and a semicolon for the separator, which will require "csv2".}

\item{AttributesDescription}{A vector with information about each of the attributes added.}

\item{Mode}{Character (string), indicating the name for the mode of the network.}
}
\value{
A "dataproject" object containing the nodal attribute data.
}
\description{
Takes nodal attribute data and adds these to the "$Attributes" part of an existing "dataproject".
\itemize{
\item Last updated: 1 May 2022.
}
}
\examples{
#' ## We can consider an existing network and add an attribute
## Create a network with ordinal categories (0='no relation', 1='acquaintance', 2='friend',
##    3='good friend') as an (adjacency) matrix:
Friendship3<-matrix(c(0,1,1, 2,0,3, 0,2,0),3,3, byrow=TRUE)
## Add node names to the matrix:
rownames(Friendship3)<-c("a","b","c")
colnames(Friendship3)<-rownames(Friendship3)

## Now create the "dataproject", which we call "SchoolClass_Project1"
SchoolClass_Project1<-xCreateProject(GeneralDescription="Dataset among 3 students",
     NetworkName="Friendship",
     NETFILE1=Friendship3,
     FileType="Robject",
     InFormatType="AdjMat",
     NetworkDescription="Strength (0=no relation, 1=acquaintance, 2=friend, 3=good friend)",
     Mode=c("Students"),
     Directed=TRUE,
     Loops=FALSE,
     Values="Ordinal",
     Class="matrix",
     References="No references")
## Check the newly created "dataproject" object
SchoolClass_Project1

## Add nodal attribute data to an existing example "dataproject":
Age3<-c(32,65,24)
Gender3<-c(1,2,1)
X1<-data.frame(Age3,Gender3)
rownames(X1)<-c("a","b","c")
SchoolClass_Project1<-xAddAttributesToProject(SchoolClass_Project1, ATTFILE1=X1, FileType="Robject",
      AttributesDescription=c("Age in years","Gender, where 2=Male and 1=Female"), Mode="Students")
## Check that the attribute data have been added:
SchoolClass_Project1
}
\author{
Filip Agneessens, \email{filipagneessens2@gmail.com}
}
