% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeLevelMeasures.R
\name{xAlterCompositionCat}
\alias{xAlterCompositionCat}
\title{Information based on ego's alters values on a categorical variable.}
\usage{
xAlterCompositionCat(
  NET1,
  ATTR1,
  Measures = c("Count", "Prop", "BlauH", "IQV"),
  Loops = FALSE,
  Mode = "OneMode"
)
}
\arguments{
\item{NET1}{A network dataset stored as an object of class "matrix".
A one-mode network with (n) nodes, or a two-mode network
with (m) nodes of mode "A" (rows) and (n) nodes of mode "B" (columns).}

\item{ATTR1}{A categorical (nominal) attribute vector of size (n), which
contains (k) unique categories.}

\item{Measures}{The choice of output to be provided. Choices are:
\itemize{
\item "Count", which provides as many columns as there are unique
categories (k) in the nominal variable (\code{ATTR1}), and gives the number of times
ego is connected with alters with a specific value using ("F.1","F.2",...).
\item "Prop", which provides as many columns as there are unique
categories (k) in the nominal variable (\code{ATTR1}), and gives the proportion of
times ego is connected with alters with a specific value using ("P.1","P.2",
...).
\item "BlauH", Blau's measure of heterogeneity, is 1 minus the sum of the
squares of the proportions of each value of the categorical variable in ego's
network.
\item "IQV", provides the Index of Qualitative Variation, which is a normalized
version of the Blau index, which is the value for the Blau index divided by 1-1/k.
}}

\item{Loops}{Whether to ignore the Loops (default is to ignore Loops).}

\item{Mode}{Whether NET1 is a one-mode ("OneMode") or two-mode ("TwoMode") network.
The default is "Autodetect", which will detect whether the network is a one-mode
or a two-mode network based on the information provided, or based on additional
information linked to the dataset.}
}
\value{
A matrix containing different measures (columns) for each node (row).
}
\description{
This function provide for each ego information about the (relative) presence of alters
with specific values on an categorical (nominal) variable and calculates measures of
heterogeneity based on these values.
}
\details{
The input is a one-mode or two-mode network (\code{NET1}) and a vector containing
a categorical (nominal) variable with a (limited) set of values (\code{ATTR1}).
It extracts for each node (ego) the nominal values for the alters that
ego is directed connected with, and based on this information, provides the
frequencies with which the alters of ego belong to each of the (k)
unique categorical value.
Based on the frequency it then calculates a measure of heterogeinity (IQV - Index of
Qualitative Variation).

The output contain a table with the frequencies for each categorical value and a
measure of heterogeneity for the values of a categorical variable among each ego's alters.

As an output it provides by default:
\itemize{
\item "Valid": the number of cases for which a valid outgoing connections ego has,
\item "EgoValue": what the nominal value for ego is (only possible for one-mode network)
\item "Degree": what the "binary" outdegree is for ego, i.e. how many unique outgoing connections ego has: "Degree"
\item "W.Degr": the sum of the outgoing ties weighted by their strength.
(For binary network data this is the same as "Degree").
\item "F.1", "F.2", ... k columns, where each columns gives the number of times ego is connected with alters with a specific value ("1","2",...)
\item "P.1", "P.2", ... k columns, where each columns gives the proportion of ego's alters that belong to category "1", "2", ...
\item "BlauH": Blau's measure of heterogeneity, is 1 minus the sum of the squares of the proportions of each value of the categorical variable in ego's network.
A person connected to equal numbers of sociologists, psychologists and economists, will have a Heterogeneity measure of 0.333,  calculated as 1 - ( (1/3)^2 + (1/3)^2) + (1/3)^2) ).
\deqn{ BlauH = 1-\sum(p[k]^2)}
\item "IQV": Index of Qualitative Variation is a normalized version of this index and is equal to the previous column divided by 1-1/k.
Using:
\code{\link{xCreateProject}}: function in this package.
}
}
\examples{
## Using Krackhardt's friendship network among high-tech managers with department as
## categorical attribute.
xAlterCompositionCat(Krackhardt_HighTech$Friendship, Krackhardt_HighTech$Attributes$Department)
}
