% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeLevelMeasures.R
\name{xAlterCompositionCon}
\alias{xAlterCompositionCon}
\title{Information based on ego's alters values on a continuous variable.}
\usage{
xAlterCompositionCon(
  NET1,
  ATTR1,
  Measures = c("SumAlterV", "AvAlterV", "WAvAlterV", "MinAlterV", "MaxAlterV",
    "RangeAlterV", "SDAlterV"),
  Loops = FALSE,
  Exclude0 = FALSE,
  Mode = "OneMode"
)
}
\arguments{
\item{NET1}{A network dataset stored as an object of class "matrix".
A one-mode network with (n) nodes, or a two-mode network
with (m) nodes of mode "A" (rows) and (n) nodes of mode "B" (columns).}

\item{ATTR1}{A continuous attribute vector of size (n).}

\item{Measures}{The choice of output to be provided. See description.}

\item{Loops}{Whether to ignore the Loops (default is to ignore the Loops).}

\item{Exclude0}{Whether to exclude values 0 from the analysis (default is FALSE).}

\item{Mode}{Whether NET1 is a one-mode ("OneMode") or two-mode ("TwoMode") network.}
}
\value{
A matrix containing different measures (columns) for each node (row).
}
\description{
This function takes a one-mode or two-mode matrix (\code{NET1}) and a vector containing a continuous nodal attribute (\code{ATTR1}).
It returns for each node in the rows of the matrix information related to the attribute data of this node's alters.
For valued networks it weights the impact of alters by the strength of its ties.
For a two-mode network, results are provided for the row nodes based on attribute data of the alters (column nodes).
In this case the attribute needs to be a vector containing information for the column nodes.
As an output it provides by default:
\itemize{
\item "SumAlterV": the (weighted) sum of the values for the alters of ego (Sj(Xij*Rj)),
\item "AvAlterV": the (weighted) sum of the values for the alters of ego and divides by the number of (non-missing) alters ((Sj(Xij*Rj))/(N-1)),
\item "WAvAlterV": the (weighted) average of the values for the alters of ego as a proportion of the strength of the ties ((Sj(Xij*Rj))/(Sj(Xij))),
\item "MinAlterV": the minimum value for the alters of ego,
\item "MaxAlterV": the maximum value for the alters of ego,
\item "RangeAlterV": the range among the values for the alters of ego,
\item "SDAlterV": the standard deviation of the values for the alters of ego,
}
}
\examples{
## Consider the Friendship network for Krackhardt's high-tech managers and the attribute tenure.
xAlterCompositionCon(Krackhardt_HighTech$Friendship, Krackhardt_HighTech$Attributes$Tenure,
                     Exclude0=TRUE)
}
