% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xBetaReachBetweennessCentrality}
\alias{xBetaReachBetweennessCentrality}
\title{Calculate Beta Reach betweenness centrality.}
\usage{
xBetaReachBetweennessCentrality(NET1, Beta = 0.5)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{Beta}{A single value for Beta (max = 1).}
}
\value{
A matrix with Beta reach betweenness centrality values.
}
\description{
Calculates the Beta Reach between centrality for a one-mode or two-mode network,
i.e., the number (proportion) of times a node is in between others, where this is weights by its length.
Beta needs to be a single value (default=.5)
Missing data should be indicated by 'NA'.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
## Undirected one-mode network
xBetaReachBetweennessCentrality(Padgett_FlorentineFamilies$Marriage, Beta=.5)
## Directed one-mode network
xBetaReachBetweennessCentrality(Krackhardt_HighTech$Advice, Beta=.3)

}
