% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCentrality.R
\name{xBetaReachCentrality}
\alias{xBetaReachCentrality}
\title{Calculate Beta Reach centrality.}
\usage{
xBetaReachCentrality(NET1, Beta = 0.5, Diagonal = FALSE, Normalize = FALSE)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{Beta}{A single valsue or vector contain Beta values.}

\item{Diagonal}{Logical statement, whether diagonal should be included (default=FALSE).}

\item{Normalize}{Whether to return the normalized value (divide by theoretical maximum value) or raw counts.}
}
\value{
A matrix with Beta reach centrality values for each value of beta
}
\description{
Calculates the Beta Reach centrality for a one-mode or two-mode network,
i.e., the number/proportion of nodes reachable in k steps weighted by Beta.
Beta can be a single value or set of values.(default=.5)
Beta is expected to be between 0 and 1
The diagonal for the geodesic distance matrix is by default set to NA.
Missing data should be indicated by 'NA'.
\itemize{
\item Last updated: 17 Sep 2022.
}
}
\examples{
## Undirected one-mode network
xBetaReachCentrality(Padgett_FlorentineFamilies$Marriage, Beta=c(.5,.3))
## Directed one-mode network (out-beta reach followed by in-beta reach)
xBetaReachCentrality((Krackhardt_HighTech$Advice), Beta=c(.4), Normalize=TRUE)
xBetaReachCentrality(t(Krackhardt_HighTech$Advice), Beta=c(.4), Normalize=TRUE)

}
