% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xBetweennessCentrality}
\alias{xBetweennessCentrality}
\title{Calculates betweenness centrality.}
\usage{
xBetweennessCentrality(NET1)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}
}
\value{
A matrix with raw and normalized betweenness centrality values.
}
\description{
Calculates betweenness centrality for a one-mode or two-mode network,
i.e., the number (proportion) of times a node 'i' is in between others.
For directed networks is consider the "in-out" connections around node i,
i.e., j->i->k.
Note: Undirected counts are counted in both directions j->i->k and k->i->j.
Provide both non-normalized and normalized measure (divided by the theoretical maximum).
No missing data are allowed.
}
\examples{
## Undirected one-mode network
xBetweennessCentrality(Padgett_FlorentineFamilies$Marriage)

## Directed one-mode network
xBetweennessCentrality(Krackhardt_HighTech$Advice)

}
