% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquivalenceRoles.R
\name{xBlockmodelOptimizing}
\alias{xBlockmodelOptimizing}
\title{Finds optimal blockmodel}
\usage{
xBlockmodelOptimizing(NET1, NOG, BlockTypes, Options, NTRIALS = 50)
}
\arguments{
\item{NET1}{A network stored as a matrix}

\item{NOG}{Number of groups/clusters to consider.}

\item{BlockTypes}{Types of blocks allowed. See blockmodeling::optRandomParC.}

\item{Options}{Whether Binary or Ratio.}

\item{NTRIALS}{Number of trials considered to find the optimal solution.}
}
\value{
Vector of different groupings based on optimal solutions.
}
\description{
Takes a network and finds the optimal solution(s) given the options requested.
}
\examples{
NET1<-matrix(c(0,1,0,1,0,
               0,0,0,0,1,
               1,1,0,1,1,
               0,0,0,0,1,
               1,1,1,0,0),5,5)
xBlockmodelOptimizing(NET1,NOG=2,BlockTypes=c("nul","com"),Options="Binary")
}
\references{
Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022).
Analyzing Social Networks Using R. SAGE.
Doreian, P., Batagelj, V., & Ferligoj, A. (2005). Generalized blockmodeling, (Structural analysis in the social sciences, 25). Cambridge: Cambridge University Press.
Žiberna, A. (2007). Generalized Blockmodeling of Valued Networks. Social Networks, 29(1), 105-126. doi: 10.1016/j.socnet.2006.04.002
Žiberna, A. (2008). Direct and indirect approaches to blockmodeling of valued networks in terms of regular equivalence. Journal of Mathematical Sociology, 32(1), 57-84. doi: 10.1080/00222500701790207
Žiberna, A. (2014). Blockmodeling of multilevel networks. Social Networks, 39(1), 46-61. doi: 10.1016/j.socnet.2014.04.002
}
