% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupsCommunity.R
\name{xCliquesCoMembership}
\alias{xCliquesCoMembership}
\title{Provides the clique comembership among nodes for a one-mode network}
\usage{
xCliquesCoMembership(NET1, Min = 3)
}
\arguments{
\item{NET1}{A binary, undirected one-mode network stored as a 'matrix' object.}

\item{Min}{The minimum size for cliques being considered. By default a clique has at least 3 nodes.}
}
\value{
A matrix with comembership among nodes.
}
\description{
Using an binary, undirected one-mode network (NET1), this function calculates the comembership matrix among nodes.
Clique comembership is the number of times two nodes are part of a clique.
The network should be a binary and undirected one-mode network.
}
\examples{
## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xCliquesCoMembership(Hawthorne_BankWiring$Game)
}
\references{
Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
igraph package - https://igraph.org
}
\seealso{
\code{\link[=xCliquesMembership]{xCliquesMembership()}}, \code{\link[=xCliquesOverlap]{xCliquesOverlap()}}, \code{\link[igraph:cliques]{igraph::max_cliques()}}
}
