% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupsCommunity.R
\name{xCliquesMembership}
\alias{xCliquesMembership}
\title{#--------------------------------------------------------------------
Provides the clique membership for a one-mode network}
\usage{
xCliquesMembership(NET1, Min = 3)
}
\arguments{
\item{NET1}{A binary, undirected one-mode network stored as a 'matrix' object.}

\item{Min}{The minimum size for cliques being reported. By default a clique has at least 3 nodes.}
}
\value{
A matrix with membership for each node (row) by clique (column).
}
\description{
Using an binary, undirected one-mode network (NET1), this function gives the maximum clique solutions.
Clique membership is captured by a matrix, with nodes as rows and each clique as a column.
The minimum size of the cliques being considered are defined in (Min).
The network should be a binary and undirected one-mode network.
}
\examples{
## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xCliquesMembership(Hawthorne_BankWiring$Game)
}
\references{
Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
igraph package - https://igraph.org
}
\seealso{
\code{\link[=xCliquesCoMembership]{xCliquesCoMembership()}}, \code{\link[=xCliquesOverlap]{xCliquesOverlap()}}, \code{\link[igraph:cliques]{igraph::max_cliques()}}
}
