% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xClosenessCentrality}
\alias{xClosenessCentrality}
\title{Calculates 3 versions of closeness centrality}
\usage{
xClosenessCentrality(NET1)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}
}
\value{
A matrix with closeness centrality values.
}
\description{
Calculates the closeness centrality for a one-mode or two-mode network.
Three versions of closeness are provided: Freeman's closeness, reciprocal closeness
and Valente-Foreman's closeness centrality measure.
Each version also is normalized.
For directed networks it calculates the out-closeness.
To calculate the in-closeness simply transpose the matrix by using t().
No missing data are allowed. Loops for the geodesic distance is set to 0.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
## Undirected one-mode network
xClosenessCentrality(Padgett_FlorentineFamilies$Marriage)

## Directed one-mode network (out-closeness followed by in-closeness)
xClosenessCentrality(Krackhardt_HighTech$Advice)
xClosenessCentrality(t(Krackhardt_HighTech$Advice))

}
