% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xCombineTies}
\alias{xCombineTies}
\title{Combine two networks}
\usage{
xCombineTies(NET1, NET2, Type = "Max")
}
\arguments{
\item{NET1}{The first network to be used.}

\item{NET2}{The second network to be used.}

\item{Type}{The way to combine two matrices. The six options are: "Min" (take the minimum of both ties), "Max" (take the maximum of both ties - the default),
"Av" (take the average of both ties), "Sum" (take the sum of both ties), "Prod" (take the product of both ties), "Unique" (creates a unique code for both - only to be used for matrices with limited number of values).}
}
\value{
The combined network.
}
\description{
This function allows different forms of combining two networks.
}
\examples{
M1<-matrix(c(1,2,3,4,5,6,7,8,NA),3,3)
M2<-matrix(c(3,2,1,4,5,6,7,8,NA),3,3)
xCombineTies(M1,M2)
}
