% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeNetworks.R
\name{xCompactness}
\alias{xCompactness}
\title{Calculates the compactness (breadth) of a network or for subparts of a network.}
\usage{
xCompactness(NET1, ROWS = "All", COLS = "All", Option = "Mean")
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{ROWS}{Whether to calculate the sum or average of the reciprocal distances over all nodes in the row (="All"), whether to keep each node in the row as a unique entity ("Unique"), or whether to aggregate based on some attribute variable (in which case (ROWS= name of the attribute dataset).}

\item{COLS}{Same as for rows, but here for columns. Note that ROWS and COLS can contain a different attribute variable, for example to see whether there is any difference in male or female students in their reciprocal distance to high achieving versus low achieving students (here gender would be ROWS and academic achievement would be COLS).}

\item{Option}{Whether to obtain the average ("Mean") or sum ("Sum") for the reciprocal distance matrix.}
}
\value{
A single value or matrix calculating the sum or average of compactness across all nodes, or split by category (or node).
}
\description{
Using a one-mode or two-mode binary network, this function calculates the compactness (or its inverse the breadth of a network).
Compactness is based on the reciprocal of the geodesic distance between pairs of nodes, so that two nodes at distance 1 have a reciprocal value of 1/1, two nodes at distance 2, a value of 1/2, etc.
Nodes that cannot reach each other are defined as having a distance infinity, which means the pair will obtain a value 1/inf=0 in the reciprocal distance matrix.
In practice it takes this reciprocal distance and takes the average (Option="Mean") or the sum (Option="Sum") of the values for all pairs in the network.
\itemize{
\item For a two-mode network, the result can be split into three parts (within mode A, within mode B, and between mode A and mode B nodes), and therefore we will obtain a matrix with within and between values.
\item Missing data NA are not allowed.
\item In addition, if an attribute file is specified for rows (ROWS) and/or the columns (COLS), then the average or sum is provided for each unique group in that attribute file. This might be useful to examine whether specific types of actors tend on average to be at shorter distances from specific other nodes. Note that ROWS and COLS does not have to contain the same attribute.
\item If ROWS or COLS is set to "Unique" (default is "All"), then the matrix is not aggregated for that row or column, and so the reciprocal distance to other (sets of) nodes is obtained for each node. This might be useful if you want to know, for example, if specific nodes are at shorter distance from nodes with specific types of actors.
}
}
\examples{
## Two simple undirected one-mode networks (with the same value for connectedness)
xCompactness(ASNR_Fig10x3C)
xCompactness(ASNR_Fig10x3D)
## Adding an attribute for rows and columns:
xCompactness(ASNR_Fig10x3D,ROWS=c(1,1,1,2,2,2),COLS=c(1,1,1,2,2,2))

## Some more complex examples with a two-mode network
NET2<-matrix(c(0,1,0,0,0,0,0, 1,0,0,0,0,0,0, 1,1,1,0,0,1,0, 0,1,1,0,0,0,0,
               0,0,0,1,1,0,0, 1,0,0,0,0,1,0, 0,1,1,1,0,1,0, 0,0,1,0,0,0,0),8,7,byrow=TRUE)
## When running the function on the twomode, note the use of "A" and "B" to separate
## mode A (row) from mode B (column) nodes:
xCompactness(NET2)
## When adding an attribute for ROWS, note that mode A nodes are now split further
## into 2 groups: A_1, A_2 and A_3.
xCompactness(NET2,ROWS=c(1,1,1,2,2,2,3,3))
## We can also keep the values separate for each unique node in the column:
xCompactness(NET2,ROWS=c(1,1,1,2,2,2,3,3),COLS="Unique")
## Or have a different attribute for mode A and mode B:
xCompactness(NET2,ROWS=c(1,1,1,2,2,2,3,3),COLS=c("P","E","T","T","P","T","E"))
## Or combine unique (in the rows) with aggregating all (in the columns):
xCompactness(NET2,ROWS="Unique")

}
\references{
Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
Carter T. Butts (2020). sna: Tools for Social Network Analysis. R package version 2.6. http://CRAN.R-project.org/package=sna
}
\seealso{
\code{\link[=xDensity]{xDensity()}}, \code{\link[=xConnectedness]{xConnectedness()}}, \code{\link[sna:geodist]{sna::geodist()}}
}
