% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManagement.R
\name{xCreateProject}
\alias{xCreateProject}
\title{Creating a "dataproject" containing network data}
\usage{
xCreateProject(
  GeneralDescription = NULL,
  NetworkName = NULL,
  NETFILE1,
  FileType = NULL,
  InFormatType = "AdjMat",
  NetworkDescription = NULL,
  Mode = NA,
  Directed = TRUE,
  Loops = FALSE,
  Values = NA,
  Class = "matrix",
  References
)
}
\arguments{
\item{GeneralDescription}{Detailed information about the project as a whole.}

\item{NetworkName}{The name to be given to the network.}

\item{NETFILE1}{A single social network file. This could be a csv file or an R object. The input type is defined in "FileType", while the format of the input data is defined in "InFormatType".}

\item{FileType}{The type of data that is being imported. Default is "csv". Other options are "Robject" when the data is already an object in R, and "csv2". Note that by default a csv file use a point (".") for the decimal point and a comma for the separator in most countries. However, in some Western European countries a comma is used for the decimal point and a semicolon for the separator, which will require "csv2".}

\item{InFormatType}{The format used to store the network data in a csv file. Currently only reads in an adjacency matrix "AdjMat".}

\item{NetworkDescription}{Detailed information about the network being imported.}

\item{Mode}{Character (string), indicating a name for the mode of the network. For two-mode networks this would need to contain the names for both modes. If it is a one-mode network, and Mode=NA then the mode name "A" will be given.}

\item{Directed}{Logical, whether the network should be considered directed or not. The default is TRUE.}

\item{Loops}{Logical, whether the network's self-nominations (loops) should be considered. By default this is FALSE and loops are not considered.}

\item{Values}{Character string, indicating what type of values the network contains (e.g., "Binary","Nominal","Ordinal","Rank","Interval","Ratio").}

\item{Class}{The way the data is being stored. Currently the network will be stored as an adjacency matrix in an object of class "matrix".}

\item{References}{Any references.}
}
\value{
A "dataproject" object containing network data.
}
\description{
This function creates a new "dataproject" list object, and adds the network (NETFILE1) and extra information about the network.
\itemize{
\item Last updated: 28 April 2022.
}
}
\examples{
## Create a network with ordinal categories (0='no relation', 1='acquaintance', 2='friend',
##    3='good friend') as an (adjacency) matrix:
Friendship3<-matrix(c(0,1,1, 2,0,3, 0,2,0),3,3, byrow=TRUE)
## Add node names to the matrix:
rownames(Friendship3)<-c("a","b","c")
colnames(Friendship3)<-rownames(Friendship3)

## Now create the "dataproject", which we call "SchoolClass_Project1"
SchoolClass_Project1<-xCreateProject(GeneralDescription="Dataset among 3 students",
     NetworkName="Friendship",
     NETFILE1=Friendship3,
     FileType="Robject",
     InFormatType="AdjMat",
     NetworkDescription="Strength (0=no relation, 1=acquaintance, 2=friend, 3=good friend)",
     Mode=c("Students"),
     Directed=TRUE,
     Loops=FALSE,
     Values="Ordinal",
     Class="matrix",
     References="No references")
## Check the newly created "dataproject" object
SchoolClass_Project1

}
\author{
Filip Agneessens, \email{filipagneessens2@gmail.com}
}
