% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xCentrality.R
\name{xDegreeCentrality}
\alias{xDegreeCentrality}
\title{Calculates the outdegree for a network (i.e. for each row).}
\usage{
xDegreeCentrality(NET1, Diagonal = FALSE)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{Diagonal}{Logical statement, whether diagonal should be included.}
}
\value{
A matrix with degree and normalized degree.
}
\description{
Calculates the (out)degree centrality for a one-mode or two-mode matrix.
To calculate the indegree simply transpose the matrix by using t().
Values can be binary or valued. For valued networks, the normalized
(out)degree ('nDegree') is the average value of the outgoing ties.
Any missing data (indicated by values 'NA') are ignored.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
## Directed valued one-mode with missing data and with diagonal
M1<-matrix(c(6,2,4,6,1,NA,2,NA,3),3,3, byrow=TRUE)
xDegreeCentrality(M1, Diagonal=FALSE)

## To obtain the indegree, use the transposed t()
xDegreeCentrality(t(M1), Diagonal=FALSE)

## Undirected one-mode network
xDegreeCentrality(ASNR_Fig09x1)

## Directed one-mode network
xDegreeCentrality(ASNR_Fig09x4)

## Undirected two-mode network
xDegreeCentrality(Davis_SouthernWomen$Attendance)
xDegreeCentrality(t(Davis_SouthernWomen$Attendance))

}
