% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoMode.R
\name{xDualCorePeriphery}
\alias{xDualCorePeriphery}
\title{Performs a dual core-periphery analysis.}
\usage{
xDualCorePeriphery(NET2)
}
\arguments{
\item{NET2}{A single two-mode network stored in an R object as a matrix.}
}
\value{
A membership
}
\description{
Performs a separate core-periphery analysis on row and column nodes using one-mode projections.
}
\examples{
xDualCorePeriphery(Davis_SouthernWomen$Attendance)
}
