% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoMode.R
\name{xDualStructuralEquivalence}
\alias{xDualStructuralEquivalence}
\title{Performs a dual structural equivalence analysis.}
\usage{
xDualStructuralEquivalence(NET2)
}
\arguments{
\item{NET2}{A single two-mode network stored in an R object as a matrix.}
}
\value{
A membership
}
\description{
Performs a separate structural equivalence analysis on row and column nodes using one-mode projections.
}
\examples{
xDualStructuralEquivalence(Davis_SouthernWomen$Attendance)
}
