% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeLevelMeasures.R
\name{xEgoAlterSimilarityCat}
\alias{xEgoAlterSimilarityCat}
\title{Information based the similarity/difference on a categorical between ego and the alters that ego is directly connected with.}
\usage{
xEgoAlterSimilarityCat(
  NET1,
  ATTR1,
  Measures = c("PctSame", "EIIndex", "OddsRatio", "LogOddsRatio", "YulesQ"),
  Loops = TRUE
)
}
\arguments{
\item{NET1}{A network dataset stored as an object of class "matrix".
A one-mode network with (n) nodes, or a two-mode network
with (m) nodes of mode "A" (rows) and (n) nodes of mode "B" (columns).}

\item{ATTR1}{A continuous attribute vector of size (n).}

\item{Measures}{The choice of output to be provided. See description.}

\item{Loops}{Whether to ignore the Loops (default is to ignore the Loops).}
}
\value{
A matrix containing different measures (columns) for each node (row).
}
\description{
This function takes a one-mode network (\code{NET1}) and a vector containing a categorical nodal attribute (\code{ATTR1}).
It returns for each node in the rows of the matrix information related to the similarity or difference in attribute data between the node and its alters.
As an output it provides by default:
-"a": the number of connected alters that belong to the same category as ego.
-"b": the number of connected alters that belong to a different category than ego.
-"c": the number of alters that ego is not connected to, that belong to the same category as ego.
-"b": the number of alters that ego is not connected to, that belong to a different category than ego.
}
\examples{
## Consider the Friendship network for Krackhardt's high-tech managers and the attribute age.
xEgoAlterSimilarityCat(Krackhardt_HighTech$Friendship,Krackhardt_HighTech$Attributes$Department)
}
