% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xEigenvectorCentrality}
\alias{xEigenvectorCentrality}
\title{Calculates the eigenvector centrality for a network (i.e. for each row).}
\usage{
xEigenvectorCentrality(NET1, Loops = FALSE)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{Loops}{Logical statement, whether Loops should be used.}
}
\value{
A matrix with eigenvector centrality and normalized eigenvector centrality (max=1).
}
\description{
Calculates the eigenvector centrality for a one-mode network.
Tends to be 'problematic' for directed networks.
Values can be binary or valued. For valued networks.
The normalized version ensures that the maximum value is 1.
No missing data are allowed.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
##Undirected networks
xEigenvectorCentrality(ASNR_Fig09x1)
xEigenvectorCentrality(ASNR_Fig09x2)
}
