% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xGeodesicDistance}
\alias{xGeodesicDistance}
\title{Calculates the geodesic distance for a network}
\usage{
xGeodesicDistance(NET1)
}
\arguments{
\item{NET1}{The network to be used.}
}
\value{
The geodesic distance for a network.
}
\description{
This function calculates the geodesic distance for a networks.
}
\examples{
M1<-matrix(c(0,NA,1,1,0,0,0,NA,0),3,3)
xGeodesicDistance(M1)
}
