% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeNetworks.R
\name{xGroupHomophily}
\alias{xGroupHomophily}
\title{Calculates measures of homophily for a one-mode network.}
\usage{
xGroupHomophily(NET1, ATTR1)
}
\arguments{
\item{NET1}{A one-mode binary network stored as a 'matrix' object.}

\item{ATTR1}{A vector containing the attribute.}
}
\value{
A vector with measures related to homophily.
}
\description{
Using a one-mode binary network, this function calculates measures of homophily.
This includes the proportion homophilous ties, the EI index and Yule's Q.
Missing data are allowed. The Loops is ignored.
Provides output:
-"a": the number of pairs of nodes that have a tie and where both nodes belong to the same category.
-"b": the number of pairs of nodes that have a tie and where both nodes belong to a different category.
-"c": the number of pairs of nodes that do not have a tie and where both nodes belong to the same category.
-"d": the number of pairs of nodes that do not have a tie and where both nodes belong to a different category.
}
\examples{
## Two examples of directed networks (See chapter 10 of Borgatti et al., 2022):
xGroupHomophily(Krackhardt_HighTech$Advice,Krackhardt_HighTech$Attributes$Department)
xGroupHomophily(Krackhardt_HighTech$Friendship,Krackhardt_HighTech$Attributes$Department)

}
\references{
Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
}
\seealso{
\code{\link[=xDensity]{xDensity()}}, \code{\link[=xEgoAlterSimilarityCat]{xEgoAlterSimilarityCat()}}
}
