% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateMethods.R
\name{xHierarchicalClustering}
\alias{xHierarchicalClustering}
\title{Provides the hierarchical clustering solution for a square matrix}
\usage{
xHierarchicalClustering(
  NET1,
  Input = NULL,
  Method = "complete",
  TitleDendrogram = "Cluster Dendrogram",
  NOC = NULL
)
}
\arguments{
\item{NET1}{A binary, undirected one-mode network stored as a 'matrix' object.}

\item{Input}{Whether the input matrix contains similarities ("Similarities") or differences ("Differences")}

\item{Method}{which clustering method to use. Default is "complete". See the hclust function.}

\item{TitleDendrogram}{The title to be used in the plot}

\item{NOC}{If containing a single value or vector, cluster membership is provided as output.}
}
\value{
A matrix where each row and column are unique cliques found, and the values are the number of nodes two cliques have in common.
}
\description{
Using a square matrix, representing either similarities or differences, this function calculates the clustering solution.
This function relies on hclust and plot, and therefore incorporates their respective options. If requested (by adding values to NOC),
cluster membership is offered for a solution with k clusters.
}
\examples{
## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
CCM<-xCliquesCoMembership(ASNR_Fig11x1,Min=3)
xHierarchicalClustering(CCM, Input="Similarities", Method="single")
MATY<-matrix(c(2,3,1,
               3,5,6,
               1,6,4),3,3)
xHierarchicalClustering(MATY, Input="Similarities")
}
\references{
Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
}
\seealso{
\code{\link[=xCliquesMembership]{xCliquesMembership()}}, \code{\link[=xCliquesMembership]{xCliquesMembership()}}, \code{\link[=hclust]{hclust()}}
}
