% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xImputeMissingData}
\alias{xImputeMissingData}
\title{Impute missing data}
\usage{
xImputeMissingData(NET1, Type = "Density", Loops = FALSE)
}
\arguments{
\item{NET1}{The network to be used.}

\item{Type}{The way to impute missing data. The current option is: "Density" (take the observed density as a basis).}

\item{Loops}{Whether to include self-nominations.}
}
\value{
An imputed network.
}
\description{
This function imputes missing data in a networks.
}
\examples{
M1<-matrix(c(0,NA,1,1,0,0,0,NA,0),3,3)
xImputeMissingData(M1, Type="Density")
}
