% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xInducedCentrality}
\alias{xInducedCentrality}
\title{Calculate induced centrality.}
\usage{
xInducedCentrality(NET1)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}
}
\value{
A matrix with a series of induced centrality measures, followed by a split between endogeneous and exogeneous.
}
\description{
Induced centrality calculates the change in specific network properties (graph invariants) when a node is removed.
A graph invariant can be used to induce a centrality score by measuring the contribution each node makes to the network property (e.g. the number of components).
The induced centrality of a node is the value for the graph invariant for the observed network minus the value for the graph invariant for the observed network but with the node deleted.
This routine calculates the induced centrality for the following network properties:
\itemize{
\item "kReach1": Number of nodes reachable within geodesic distance 1 (Degree)
\item "kReach2": Number of nodes reachable within geodesic distance 2
\item "kReach3": Number of nodes reachable within geodesic distance 3
\item "kReach4": Number of nodes reachable within geodesic distance 4
\item "kReachAll": Number of nodes reachable within a finite number of nodes
\item "RecipCloseness": Sum of reciprocal geodesic distance to all other nodes
\item "Betweenness": Number of times a node is between other nodes
No missing data are allowed.
\item Last updated: 1 February 2022.
}
}
\examples{
## Undirected one-mode network
xInducedCentrality(Padgett_FlorentineFamilies$Marriage)

## Directed one-mode network
xInducedCentrality(Krackhardt_HighTech$Advice)
xInducedCentrality(t(Krackhardt_HighTech$Advice))
}
