% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xNegativeDegreeCentrality}
\alias{xNegativeDegreeCentrality}
\title{Calculates the outdegree for negative relations (i.e. for each row).}
\usage{
xNegativeDegreeCentrality(NET1N)
}
\arguments{
\item{NET1N}{A one-mode or two-mode negative network stored as a 'matrix' object.}
}
\value{
A matrix with outdegree centrality measures.
}
\description{
Calculates the (out)degree centrality for a one-mode containing a negative relation.
To calculate the indegree simply transpose the matrix by using t().
This function calculates degree centrality for negative relations as explained in Everett and Borgatti (2014).
Only for binary data.
Any missing data (indicated by values 'NA') are ignored.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
## Undirected one-mode network # TABLE 1
xNegativeDegreeCentrality(Read_NewGuinea$Opposition)

## Directed one-mode network # TABLE 3
NEG1<-((Sampson_Monastery$Disesteem+Sampson_Monastery$Dislike+Sampson_Monastery$Blame
        +Sampson_Monastery$NegativeInfluence)>0)*1
xNegativeDegreeCentrality(NEG1)
xNegativeDegreeCentrality(t(NEG1))

}
\references{
Everett, M.G. Borgatti, S.P. (2014) Networks containing negative ties. Social Networks 38, 111-120.
}
