% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xNegativeWeightedCentrality}
\alias{xNegativeWeightedCentrality}
\title{Calculates the weighted centrality for negative relations (i.e. for each row).}
\usage{
xNegativeWeightedCentrality(NET1N)
}
\arguments{
\item{NET1N}{A one-mode negative network stored as a 'matrix' object.}
}
\value{
A matrix with weighted centrality measures.
}
\description{
This function calculates a weighted centrality for negative relations as explained in Everett and Borgatti (2014).
No missing data allowed.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
## Undirected one-mode network # TABLE 1
xNegativeWeightedCentrality(Read_NewGuinea$Opposition)

## Directed one-mode network # TABLE 3
NEG1<-((Sampson_Monastery$Disesteem+Sampson_Monastery$Dislike+Sampson_Monastery$Blame
        +Sampson_Monastery$NegativeInfluence)>0)*1
xNegativeWeightedCentrality(NEG1)
xNegativeWeightedCentrality(t(NEG1))

}
\references{
Everett, M.G. Borgatti, S.P. (2014) Networks containing negative ties. Social Networks 38, 111-120.
}
