% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-xucinet.R
\name{xnrelations}
\alias{xnrelations}
\title{Number of relations in a dataset}
\usage{
xnrelations(net)
}
\arguments{
\item{net}{A network (any accepted form).}
}
\value{
An integer.
}
\description{
Number of relations in a dataset
}
