% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{xPermuteQAP}
\alias{xPermuteQAP}
\title{Takes a one-mode network and randomly permutes both rows and columns}
\usage{
xPermuteQAP(NET0)
}
\arguments{
\item{NET0}{A network (currently only adjacency matrices are supported).}
}
\value{
A permuted adjacency matrix, where rows and columns are reodered in the same way.
}
\description{
This function takes a single one-mode network and randomly reorders (permutes) both rows and columns. This function is used for QAP-based analysis.
\itemize{
\item Last updated: 23 April 2022.
}
}
\examples{
MAT1<-matrix(c(2,4,2,3,3,1,3,4,1),3,3)
xPermuteQAP(MAT1)
}
\author{
Filip Agneessens \href{mailto:filipagneessens2@gmail.com}{filipagneessens2@gmail.com}
}
