% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{xQAPRegression}
\alias{xQAPRegression}
\title{Performs a regression analysis using a classic (and a permutation-based) QAP significance test on a network variable}
\usage{
xQAPRegression(
  NET1,
  LIST2,
  NPERM = 1000,
  Directed = TRUE,
  Method = "qapspp",
  Loops = FALSE
)
}
\arguments{
\item{NET1}{A first network (adjacency matrix), capturing the values for the dependent variable.}

\item{LIST2}{A list object containing the networks (adjacency matrices) which will be the independent variables.}

\item{NPERM}{A positive integer value indicating the number of permutations to be used for the statistical test. If NPERM is set to NULL, NA or 0, no permutation-based significance test is performed.}

\item{Directed}{Whether the analysis should consider ties in both directions (node i to j and j to i) as separate cases. Default is TRUE.}

\item{Method}{The type of permutation being used. The options are the Double Dekker semi-partialling ("qapspp"), permuting the dependent variable (Y) ("qapy") and independently permuting each of the independent variables (X) ("qapallx"). See the argument "nullhyp" in the netlm function for the sna packages for details.}

\item{Loops}{Whether to allow self-loops.}
}
\value{
Prints the regression output.
}
\description{
This function takes a network as a dependent variable and regresses it on a set of independent networks. The approach uses each tie as a case.
It returns the regression coefficients and the classic (two-tailed) significance tests. Because of the independence of cases (ties) the classic significance test should not be relied on.
If NPERM (number of permutations) is set to a value higher than 0 (default is 1000), a quadratic assignment permutations-based (QAP) significance is also produced,
which will provide information about the regression coefficients for the permuted variables. A higher value than 1000 is generally advisable.
Both the one-tailed ("AsSmall" or "AsLarge") and the two-tailed ("AsExtreme") test are given. This function relies on the netlm function from the sna package.
\itemize{
\item Last updated: 26 April 2022.
}
}
\examples{
MATRIX1<-matrix(c(0,4,2,3, 3,0,4,1, 4,8,0,2, 6,2,6,0),4,4)
MATRIX2<-matrix(c(0,1,2,1, 3,0,3,7, 4,7,0,2, 5,2,2,0),4,4)
MATRIX3<-matrix(c(0,4,5,3, 2,0,2,8, 4,4,0,2, 3,5,3,0),4,4)

LIST2<-list(MATRIX2,MATRIX3)
RESULTS1<-xQAPRegression(MATRIX1,LIST2,NPERM=1000)
RESULTS1
}
\references{
Chapter 14. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
Dekker, D., Krackhardt, D., and Snijders, T.A.B. (2007) Sensitivity of MRQAP tests to collinearity and autocorrelation conditions. Psychometrika 72, 563-581.
Krackhardt, D. (1988) Predicting with networks: Nonparametric multiple-regression analysis of dyadic data. Social Networks 10, 359-381.
Butts, Carter T. (2016). sna: Tools for Social Network Analysis. R package version 2.4.
Butts, Carter T. (2008). Social Network Analysis with sna. Journal of Statistical Software, 24(6).
}
\seealso{
\code{\link[=xQAPCorrelation]{xQAPCorrelation()}}
}
