% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Centrality.R
\name{xReachCentrality}
\alias{xReachCentrality}
\title{Calculate kReach centrality.}
\usage{
xReachCentrality(
  NET1,
  kReach = 2,
  Loops = FALSE,
  Normalize = FALSE,
  Cumulative = TRUE
)
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{kReach}{A single value or vector contain kReach values.}

\item{Loops}{Logical statement, whether Loops should be included (default=FALSE).}

\item{Normalize}{Whether to return the normalized value or raw counts.}

\item{Cumulative}{Whether to consider nodes reached in exactly k steps (FALSE) or in <=k steps (TRUE).}
}
\value{
A matrix with kReach centrality values.
}
\description{
Calculates the kReach centrality for a one-mode or two-mode network,
i.e., the number/proportion of nodes reachable in k steps (if Cumulative=TRUE),
or in exactly k steps if Cumulative=FALSE.
k is defined with the argument kReach, which can be a single value or a set of values.
The Loops for the geodesic distance matrix is by default set to NA.
Missing data should be indicated by 'NA'.
\itemize{
\item Last updated: 1 February 2022.
}
}
\examples{
## Undirected one-mode network
xReachCentrality(Padgett_FlorentineFamilies$Marriage, kReach=c(1:6))

## Directed one-mode network (out-reach followed by in-reach)
xReachCentrality(Krackhardt_HighTech$Advice, kReach=c(1:6))
xReachCentrality(t(Krackhardt_HighTech$Advice), kReach=c(1:6))

}
