% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoMode.R
\name{xTwoModeToBipartite}
\alias{xTwoModeToBipartite}
\title{Transforms a two-mode network into a bipartite network}
\usage{
xTwoModeToBipartite(NET2, First = "RowNodes")
}
\arguments{
\item{NET2}{A single two-mode network stored in an R object as a matrix.}

\item{First}{Whether to first put RowNodes or ColumnNodes.}
}
\value{
A one-mode bipartite network
}
\description{
This function takes a two-mode network stored as a matrix and transforms it
into a bipartite network where the row and columns consists of both actors and events.
}
\examples{
xTwoModeToBipartite(Davis_SouthernWomen$Attendance)
}
