% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases-1e.R
\name{xucinet-1e}
\alias{xucinet-1e}
\alias{xAddAttributesToProject}
\alias{xAddToProject}
\alias{xAlterCompositionCat}
\alias{xAlterCompositionCon}
\alias{xAttributeToMatrix}
\alias{xAttributeToNetwork}
\alias{xAutoRegression}
\alias{xAutoregression}
\alias{xAverageDistance}
\alias{xBetaCentrality}
\alias{xBetaReachCentrality}
\alias{xBetweennessCentrality}
\alias{xBiCliques}
\alias{xBiComembership}
\alias{xBlockmodel}
\alias{xBlockmodelOptimizing}
\alias{xBlockOptimization}
\alias{xBlockOptimize}
\alias{xBreadth}
\alias{xCentralization}
\alias{xCliqueComembership}
\alias{xCliqueOverlap}
\alias{xCliques}
\alias{xCliquesCoMembership}
\alias{xCliquesMembership}
\alias{xCliquesOverlap}
\alias{xClosenessCentrality}
\alias{xCombineNodes}
\alias{xCombineTies}
\alias{xCompactness}
\alias{xComponents}
\alias{xConcentration}
\alias{xConnectedness}
\alias{xContinuousCorePeriphery}
\alias{xConvertData}
\alias{xCorePeriphery}
\alias{xCorrelation}
\alias{xCorrespondenceAnalysis}
\alias{xCreateProject}
\alias{xCyclicality}
\alias{xDegreeCentrality}
\alias{xDensity}
\alias{xDichotomize}
\alias{xDiscreteCorePeriphery}
\alias{xDisplayBlocked}
\alias{xDualCorePeriphery}
\alias{xDualDiscreteCorePeriphery}
\alias{xDualLouvainMethod}
\alias{xDualStructuralEquivalence}
\alias{xEgoAlterSimilarityCat}
\alias{xEgoAlterSimilarityCon}
\alias{xEgonetStructure}
\alias{xEigenvectorCentrality}
\alias{xFastGreedy}
\alias{xFragmentation}
\alias{xGeodesicDistance}
\alias{xGirvanNewman}
\alias{xGroupHomophily}
\alias{xGroupsCohesion}
\alias{xHierarchicalCluster}
\alias{xHierarchicalClustering}
\alias{xImputeMissingData}
\alias{xInducedCentrality}
\alias{xLabelProp}
\alias{xLabelPropagation}
\alias{xLouvainMethod}
\alias{xLRQAP}
\alias{xMDS}
\alias{xMRQAP}
\alias{xMultipleTieComposition}
\alias{xNegativeDegreeCentrality}
\alias{xNegativeWeightedCentrality}
\alias{xNormalize}
\alias{xPermuteCor}
\alias{xPermuteCorrelation}
\alias{xPermuteQAP}
\alias{xPermuteRegression}
\alias{xPNCentrality}
\alias{xProfileSimilarity}
\alias{xProportionReach}
\alias{xQAPCorrelation}
\alias{xQAPLogisticRegression}
\alias{xQAPRegression}
\alias{xReachCentrality}
\alias{xReciprocity}
\alias{xREGE}
\alias{xRegression}
\alias{xRemoveFromProject}
\alias{xStructuralEquivalence}
\alias{xStructuralHoles}
\alias{xSymmetrize}
\alias{xTieCombine}
\alias{xTieComposition}
\alias{xTieDichotomize}
\alias{xTieNormalization}
\alias{xTieSymmetrize}
\alias{xTieTranspose}
\alias{xTransitivity}
\alias{xTranspose}
\alias{xTwoModeToBipartite}
\alias{xTwoModeToOneMode}
\alias{xTwomodeToOnemode}
\alias{xValuedTieComposition}
\alias{xWalkTrap}
\alias{xWalktrap}
\title{ASNR 1e function names}
\usage{
xAddAttributesToProject(...)

xAddToProject(...)

xAlterCompositionCat(...)

xAlterCompositionCon(...)

xAttributeToMatrix(...)

xAttributeToNetwork(...)

xAutoRegression(...)

xAutoregression(...)

xAverageDistance(...)

xBetaCentrality(...)

xBetaReachCentrality(...)

xBetweennessCentrality(...)

xBiCliques(...)

xBiComembership(...)

xBlockmodel(...)

xBlockmodelOptimizing(...)

xBlockOptimization(...)

xBlockOptimize(...)

xBreadth(...)

xCentralization(...)

xCliqueComembership(...)

xCliqueOverlap(...)

xCliques(...)

xCliquesCoMembership(...)

xCliquesMembership(...)

xCliquesOverlap(...)

xClosenessCentrality(...)

xCombineNodes(...)

xCombineTies(...)

xCompactness(...)

xComponents(...)

xConcentration(...)

xConnectedness(...)

xContinuousCorePeriphery(...)

xConvertData(...)

xCorePeriphery(...)

xCorrelation(...)

xCorrespondenceAnalysis(...)

xCreateProject(...)

xCyclicality(...)

xDegreeCentrality(...)

xDensity(...)

xDichotomize(...)

xDiscreteCorePeriphery(...)

xDisplayBlocked(...)

xDualCorePeriphery(...)

xDualDiscreteCorePeriphery(...)

xDualLouvainMethod(...)

xDualStructuralEquivalence(...)

xEgoAlterSimilarityCat(...)

xEgoAlterSimilarityCon(...)

xEgonetStructure(...)

xEigenvectorCentrality(...)

xFastGreedy(...)

xFragmentation(...)

xGeodesicDistance(...)

xGirvanNewman(...)

xGroupHomophily(...)

xGroupsCohesion(...)

xHierarchicalCluster(...)

xHierarchicalClustering(...)

xImputeMissingData(...)

xInducedCentrality(...)

xLabelProp(...)

xLabelPropagation(...)

xLouvainMethod(...)

xLRQAP(...)

xMDS(...)

xMRQAP(...)

xMultipleTieComposition(...)

xNegativeDegreeCentrality(...)

xNegativeWeightedCentrality(net, ...)

xNormalize(...)

xPermuteCor(...)

xPermuteCorrelation(...)

xPermuteQAP(...)

xPermuteRegression(...)

xPNCentrality(...)

xProfileSimilarity(...)

xProportionReach(...)

xQAPCorrelation(...)

xQAPLogisticRegression(...)

xQAPRegression(...)

xReachCentrality(...)

xReciprocity(...)

xREGE(...)

xRegression(...)

xRemoveFromProject(...)

xStructuralEquivalence(...)

xStructuralHoles(...)

xSymmetrize(...)

xTieCombine(...)

xTieComposition(...)

xTieDichotomize(...)

xTieNormalization(...)

xTieSymmetrize(...)

xTieTranspose(...)

xTransitivity(...)

xTranspose(...)

xTwoModeToBipartite(...)

xTwoModeToOneMode(...)

xTwomodeToOnemode(...)

xValuedTieComposition(...)

xWalkTrap(...)

xWalktrap(...)
}
\arguments{
\item{...}{Passed unchanged to the 2.0 function.}
}
\value{
Whatever the 2.0 function returns.
}
\description{
The function names used in \emph{Analyzing Social Networks with R} (1st edition)
and the xUCINET 0.x package, kept so that code written against the first
edition still runs. Each is a thin wrapper that names its 2.0 replacement
and calls it.
}
\details{
These are not the package's interface and are not documented in the
vignettes. New code should use the 2.0 names; \code{xucinet_1e_names()} lists
every alias and what it maps to.
}
\keyword{internal}
