% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zachary_attr}
\alias{zachary_attr}
\title{Karate club member attributes}
\format{
A data frame, 34 rows keyed by member id, 3 columns:
\describe{
\item{Support}{faction membership from a max-flow / min-cut partition}
\item{Strength}{strength of support for the chosen faction}
\item{Club}{which club the member ended up in: 1 = Mr Hi's new club,
34 = the original club}
}
}
\source{
As \link{zachary}.
}
\usage{
zachary_attr
}
\description{
Karate club member attributes
}
\examples{
table(zachary_attr$Club)
}
\seealso{
\link{zachary}
}
\keyword{datasets}
