% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zachary_KarateClub.R
\docType{data}
\name{Zachary_KarateClub}
\alias{Zachary_KarateClub}
\title{Zachary's University Karate Club}
\format{
An object of class \code{list} of length 5.
}
\source{
{UCINET}
}
\usage{
Zachary_KarateClub
}
\description{
Network data among 34 members of a university karate club collected by Wayne Zachary between 1970 and 1972. Zachary (1977) used these data and an information flow model of network conflict resolution to explain the split-up of this group following disputes among the members.
See $ProjectInfo for more information.
}
\examples{
Zachary_KarateClub

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
