% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baker_Journals.R
\docType{data}
\name{Baker_Journals}
\alias{Baker_Journals}
\title{Citation network among social work journals collected by Baker.}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Baker_Journals
}
\description{
Donald Baker collected data on the citations among 20 social work journals for 1985-1986. The data consist of the number of citations from articles in one journal to articles in another journal. Using hierarchical cluster analysis, he found that the network seems to have a core-periphery structure, and argues that American journals tend to be less likely to involve international literature.
See $ProjectInfo for more information.
}
\examples{
Baker_Journals

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
