% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bernard_HamRadio.R
\docType{data}
\name{Bernard_HamRadio}
\alias{Bernard_HamRadio}
\title{Amateur HAM radio calls by Bernard and colleagues.}
\format{
An object of class \code{list} of length 5.
}
\source{
{UCINET}
}
\usage{
Bernard_HamRadio
}
\description{
Bernard & Killworth, later with the help of Sailer, collected five sets of data on human interactions in bounded groups and on the actors' ability to recall those interactions. In each study they obtained measures of social interaction among all actors, and ranking data based on the subjects' memory of those interactions. This dataset is based on amateur HAM radio calls made over a one-month period.
See $ProjectInfo for more information.
}
\examples{
Bernard_HamRadio

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
