% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borgatti_Camp92.R
\docType{data}
\name{Borgatti_Camp92}
\alias{Borgatti_Camp92}
\title{Network among participants of a workshop}
\format{
An object of class \code{list} of length 5.
}
\source{
{UCINET}
}
\usage{
Borgatti_Camp92
}
\description{
These data were collected by Steve Borgatti, Russ Bernard, Bert Pelto and Gery Ryan at the 1992 NSF Summer Institute on Research Methods in Cultural Anthropology. This was a 3-week course given to 14 carefully selected participants. Network data were collected at the end of each week. These data were collected at the end of the second week. The data were collected by placing each person's name on a card and asking each respondent to sort the cards in order of how much interaction they had with that person since the beginning of the course (known informally as 'camp'). This results in rank order data in which a '1' indicates the most interaction while a '17' indicates the least interaction.
See $ProjectInfo for more information.
}
\examples{
Borgatti_Camp92

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
