% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borgatti_Cities.R
\docType{data}
\name{Borgatti_Cities}
\alias{Borgatti_Cities}
\title{Matrix capturing distances between US cities}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Borgatti_Cities
}
\description{
This network contain the distance in miles between some of the major cities in the US. The distances are based on how many miles it takes to drive from one city to the other.
See $ProjectInfo for more information.
}
\examples{
Borgatti_Cities

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
