% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Borgatti_Scientists504.R
\docType{data}
\name{Borgatti_Scientists504}
\alias{Borgatti_Scientists504}
\title{Collaboration among 504 scientists.}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Borgatti_Scientists504
}
\description{
This is the main component of the Borgatti_Scientists960 dataset.
See $ProjectInfo for more information.
}
\examples{
Borgatti_Scientists504

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
