% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Burkhardt_GovernmentAgency.R
\docType{data}
\name{Burkhardt_GovernmentAgency}
\alias{Burkhardt_GovernmentAgency}
\title{Advice among employees in a government agency collected by Burkhardt and Brass.}
\format{
An object of class \code{list} of length 8.
}
\source{
{UCINET}
}
\usage{
Burkhardt_GovernmentAgency
}
\description{
Network data capturing who communicated with whom about work matters at a government agency at five points in tie. The data were collected by Burkhardt and Brass on 75 employees.
See $ProjectInfo for more information.
}
\examples{
Burkhardt_GovernmentAgency

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
