% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Davis_SouthernWomen.R
\docType{data}
\name{Davis_SouthernWomen}
\alias{Davis_SouthernWomen}
\title{Two-mode network of Southern women attending events.}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Davis_SouthernWomen
}
\description{
These data were collected by Davis, Gardner and Gardner in the 1930s. They represent attendance at 14 social events by 18 Southern women. The result is a person-by-event matrix, with people in the rows and events in the columns. A cell (i,j) is 1 if person i attended social event j, and 0 otherwise. (See also Chapter 13 of Borgatti et al., 2022).
See $ProjectInfo for more information.
}
\examples{
Davis_SouthernWomen

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
