% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freeman_EIES.R
\docType{data}
\name{Freeman_EIES}
\alias{Freeman_EIES}
\title{Freeman's EIES network dataset}
\format{
An object of class \code{list} of length 6.
}
\source{
UCINET
}
\usage{
Freeman_EIES
}
\description{
This data arose from an early experiment on computer mediated communication. Fifty academics interested in interdisciplinary research were allowed to contact each other via an Electronic Information Exchange System (EIES). The data collected consisted of all messages sent as well as acquaintance relationships at two time periods (collected via a questionnaire). The data includes the 32 actors who completed the study. In addition, attribute data on primary discipline and number of citations was recorded.
See $ProjectInfo for more information.
}
\references{
See $ProjectInfo$References
}
\keyword{datasets}
