% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hawthorne_BankWiring.R
\docType{data}
\name{Hawthorne_BankWiring}
\alias{Hawthorne_BankWiring}
\title{Interactions among employees in a bank wiring room (Hawthorne plant).}
\format{
An object of class \code{list} of length 9.
}
\source{
{UCINET}
}
\usage{
Hawthorne_BankWiring
}
\description{
These are the observational data on 14 Western Electric (Hawthorne Plant) employees from the bank wiring room first presented in Roethlisberger & Dickson (1939). The data are better known through a scrutiny made of the interactions in Homans (1950), as well as the CONCOR analysis presented in Breiger et al (1975). The employees worked in a single room and include two inspectors (I1 and I3), three solderers (S1, S2 and S3), and nine wiremen or assemblers (W1 to W9). Contains six types of interactions among employees.
See $ProjectInfo for more information.
}
\examples{
Hawthorne_BankWiring

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
