% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Krackhardt_HighTech.R
\docType{data}
\name{Krackhardt_HighTech}
\alias{Krackhardt_HighTech}
\title{Network ties among managers in a high-tech company collected by Krackhardt.}
\format{
An object of class \code{list} of length 6.
}
\source{
{UCINET}
}
\usage{
Krackhardt_HighTech
}
\description{
These are data collected by David Krackhardt among the managers of a high-tech company. The company manufactured high-tech equipment on the West Coast of the United States and had just over 100 employees with 21 managers. Contains three types of network relations and four attributes.
See $ProjectInfo for more information.
}
\examples{
Krackhardt_HighTech

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
