% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazega_Lawyers.R
\docType{data}
\name{Lazega_Lawyers}
\alias{Lazega_Lawyers}
\title{Network ties among lawyers in a law firm.}
\format{
An object of class \code{list} of length 6.
}
\source{
{UCINET}
}
\usage{
Lazega_Lawyers
}
\description{
This data contain the social relations among 71 attorneys (partners and associates) in a Northeastern US corporate law firm between 1988-1991 in New England.
See $ProjectInfo for more information.
}
\examples{
Lazega_Lawyers

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
