% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainas_Terro.R
\docType{data}
\name{Mainas_Terro}
\alias{Mainas_Terro}
\title{Contacts among people collected as part of counter-terrorism/Europol.}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Mainas_Terro
}
\description{
Data obtained by Efstathios Mainas from Europol. Were collected and analyzed by both the Counter-Terrorism Unit and the Analysis Unit of Europol. Edges are the known contacts of the actors - the specific type of linkage was not revealed.
See $ProjectInfo for more information.
}
\examples{
Mainas_Terro

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
