% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Padgett_FlorentineFamilies.R
\docType{data}
\name{Padgett_FlorentineFamilies}
\alias{Padgett_FlorentineFamilies}
\title{Network among Florentine families}
\format{
An object of class \code{list} of length 5.
}
\source{
{UCINET}
}
\usage{
Padgett_FlorentineFamilies
}
\description{
This data contain the social relations among 16 Renaissance Florentine families (person aggregates) coded by John Padgett from historical documents. Breiger & Pattison (1986), in their discussion of local role analysis, use a subset of data. The two relations are business ties and marriage alliances. As Breiger & Pattison point out, the original data are symmetrically coded. This is acceptable perhaps for marital ties, but is unfortunate for the financial ties (which are almost certainly directed). Substantively, the data include families who were locked in a struggle for political control of the city of Florence in around 1430. Two factions were dominant in this struggle: one revolved around the infamous Medicis, the other around the powerful Strozzis.
See $ProjectInfo for more information.
}
\examples{
Padgett_FlorentineFamilies

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
