% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pane_Training.R
\docType{data}
\name{Pane_Training}
\alias{Pane_Training}
\title{Network ties among a training squad}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Pane_Training
}
\description{
This dataset is a trust network among members of a training squad.
See $ProjectInfo for more information.
}
\examples{
Pane_Training

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
