% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pane_Training.R
\docType{data}
\name{Pane_Training}
\alias{Pane_Training}
\title{Network ties among a training squad}
\format{
A list object (xUCINET project/dataset) containing 1 network and no attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $ProjectName, $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network: $Name, $Mode, $Directed, $Diagonal, £Values, $Class}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Pane_Training
}
\description{
This dataset is a trust network among members of a training squad.
}
\examples{
Pane_Training

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
