% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_NewGuinea.R
\docType{data}
\name{Read_NewGuinea}
\alias{Read_NewGuinea}
\title{Network ties among Gahuku-Gama sub-tribes}
\format{
An object of class \code{list} of length 5.
}
\source{
{UCINET}
}
\usage{
Read_NewGuinea
}
\description{
This data contain the social relations between Gahuku-Gama sub-tribes in NewGuinea as reported in Read (1954). Political (alliance) ties and opposition ties are reported.
See $ProjectInfo for more information.
}
\examples{
Read_NewGuinea

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
