% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rehnquist_SupremeCourt.R
\docType{data}
\name{Rehnquist_SupremeCourt}
\alias{Rehnquist_SupremeCourt}
\title{Voting among US Supreme Court Justices}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Rehnquist_SupremeCourt
}
\description{
This data contains the voting behavior among Supreme Court Justices of the United States between 1995 and 2004. From 1986 to 2005 the chief justice was William Hubbs Rehnquist. The data consist of a 2-mode matrix in which the rows are court cases and the columns are judges. A cell (i,j) = 1 if judge j voted with the majority on case i.
See $ProjectInfo for more information.
}
\examples{
Rehnquist_SupremeCourt

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
