% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Savage_TransactionFlows.R
\docType{data}
\name{Savage_TransactionFlows}
\alias{Savage_TransactionFlows}
\title{Savage's transaction flows between countries}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Savage_TransactionFlows
}
\description{
This data contain data about transaction flows between 15 countries prior to the economic crash of 1929, as reported by Savage and Deutsch in their paper in 1960.
See $ProjectInfo for more information.
}
\examples{
Savage_TransactionFlows

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
