% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Schwimmer_PapuanVillage.R
\docType{data}
\name{Schwimmer_PapuanVillage}
\alias{Schwimmer_PapuanVillage}
\title{Schwimmer's gift-giving network}
\format{
An object of class \code{list} of length 4.
}
\source{
{UCINET}
}
\usage{
Schwimmer_PapuanVillage
}
\description{
These data represent the relation of gift-giving (taro exchange) among 22 households in a Papuan village. Hage & Harary (1983) used them to illustrate a graph Hamiltonian cycle. Schwimmer points out how these ties function to define the appropriate persons to mediate the act of asking for or receiving assistance among group members. This dataset is called 'Taro' in Windows Ucinet.
See $ProjectInfo for more information.
}
\examples{
Schwimmer_PapuanVillage

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
