% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supremecourt}
\alias{supremecourt}
\title{Rehnquist court voting}
\format{
An \code{xucinet} object, 376 cases x 9 judges, 2-mode.
}
\source{
Borgatti, S. P., Everett, M. G., Johnson, J. C. and Agneessens, F.
\emph{Analyzing Social Networks}. 3rd edition. Sage.
}
\usage{
supremecourt
}
\description{
How the nine justices of the US Supreme Court voted on 376 cases between 1995
and 2004, the years of the second natural court under Chief Justice William
Rehnquist. A two-mode matrix of cases by judges.
}
\details{
A cell is 1 where the judge voted with the majority on that case and 0 where
not. Thirty-two cells across eighteen cases hold 0.5, and nine cells are
missing; UCINET stores those as 1e38 and they arrive here as \code{NA}.
}
\examples{
dim(supremecourt)
sum(is.na(as.matrix(supremecourt)))
}
\seealso{
\link{supremecourt_cases_attr}, \link{supremecourt_judges_attr}
}
\keyword{datasets}
