% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wolfe_Primates.R
\docType{data}
\name{Wolfe_Primates}
\alias{Wolfe_Primates}
\title{Wolfe's study on interaction among monkeys}
\format{
A list object (xUCINET project/dataset) containing 2 networks and 3 attributes
\describe{
\item{$ProjectInfo}{ -  Contains a series of elements and descriptions: $GeneralDescription, $AttributesDescription, $NetworksDescription, $References}
\item{$Attributes}{ -  A data.frame with attributes for nodes}
\item{$NetworkInfo}{ -  A data.frame with additional information about each network}
\item{$Strength}{ -  Network1. See $ProjectInfo$NetworksDescription for more information}
\item{$Connection}{ -  Network2. See $ProjectInfo$NetworksDescription for more information}
}
}
\source{
{UCINET}
}
\usage{
Wolfe_Primates
}
\description{
These data represent 3 months of interactions among a troop of monkeys, observed in the wild by Linda Wolfe as they sported by a river in Ocala, Florida.
}
\examples{
Wolfe_Primates

}
\references{
{See $ProjectInfo$References}
}
\keyword{datasets}
