% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoMode.R
\name{xBiCliques}
\alias{xBiCliques}
\title{Calculates the cliquemembership for a two-mode network.}
\usage{
xBiCliques(NET2, MinA = 3, MinB = 3)
}
\arguments{
\item{NET2}{A single two-mode network stored in an R object as a matrix.}

\item{MinA}{The minimum number of mode A nodes (row)}

\item{MinB}{The minimum number of mode B nodes (columns)}
}
\value{
A one-mode bipartite network
}
\description{
This function takes a two-mode network stored as a matrix and calculates the bicliques.
By default a biclique consists of at least 3 mode A and 3 mode B nodes who are all maximal connected.
}
\examples{
xBiCliques(Davis_SouthernWomen$Attendance)

}
