% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquivalenceRoles.R
\name{xCorePeriphery}
\alias{xCorePeriphery}
\title{Finds a core-periphery structure}
\usage{
xCorePeriphery(NET1)
}
\arguments{
\item{NET1}{A network stored as a matrix}
}
\value{
The eigenvector of each node, the correlation and an optimal solution.
}
\description{
Takes a network and finds the core-periphery structure using the correlation between the values and a vector with 0s and 1s (concentration).
}
\examples{
NET1<-matrix(c(0,1,0,1,0,
               0,0,0,0,1,
               1,1,0,1,1,
               0,0,0,0,1,
               1,1,1,0,0),5,5)
xCorePeriphery(NET1)
}
\references{
Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022).
Analyzing Social Networks Using R. SAGE.
}
