% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateMethods.R
\name{xCorrespondenceAnalysis}
\alias{xCorrespondenceAnalysis}
\title{Performs a correspondence analysis (CA) on a matrix}
\usage{
xCorrespondenceAnalysis(
  MAT1,
  Margins = 0.05,
  Distance = c(0, 4),
  DIM = 2,
  Color = c("Blue", "DarkRed")
)
}
\arguments{
\item{MAT1}{A rectangular matrix}

\item{Margins}{The margins added to the output figure}

\item{Distance}{The distance of the names (labels) of the nodes to the midpoint (as a \% of the overall distance).}

\item{DIM}{The number of dimensions to be considered}

\item{Color}{A vector of size 2, indicating the colors for row and column variables.}
}
\value{
A figure containing the correspondence analysis and the output from corresp().
}
\description{
Takes a rectangular matrix and performs a correspondence analysis on the matrix. This function relies on the corresp() function (MASS).
}
\examples{
MATY<-matrix(c(6,3,3,2,
               8,2,5,2,
               4,8,4,5,
               9,8,6,3,
               7,3,2,5,
               3,2,1,5,
               8,5,6,7,
               3,2,4,2),8,4,byrow=TRUE)
rownames(MATY)<-c("P1","P2","P3","P4","P5","P6","P7","P8")
colnames(MATY)<-c("V1","V2","V3","V4")
xCorrespondenceAnalysis(MATY)
}
\references{
Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022).
Analyzing Social Networks Using R. SAGE.
}
