% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeNetworks.R
\name{xDensity}
\alias{xDensity}
\title{Calculates the average or sum of all pairs of nodes in a network or for subparts of a network.}
\usage{
xDensity(NET1, Loops = FALSE, ROWS = "All", COLS = "All", Option = "Mean")
}
\arguments{
\item{NET1}{A one-mode or two-mode network stored as a 'matrix' object.}

\item{Loops}{Whether to ignore the Loops (default is to ignore Loops, i.e., Loops=FALSE).}

\item{ROWS}{Whether to calculate the sum or average over all nodes in the row (="All"), whether to keep each node in the row as a unique entity ("Unique") or whether to aggregate based on some attribute variable (in which case (ROWS= name of the attribute dataset).}

\item{COLS}{Same as for rows, but here for columns. Note that ROWS and COLS can contain a different attribute variable, for example to see whether there is any difference in male or females bully high achieving versus low achieving students (here gender would be ROWS and academic achievement would be COLS).}

\item{Option}{Whether to obtain the average ("Mean") or sum ("Sum").}
}
\value{
A single value or matrix calculating the sum or average across all nodes, or split by category (or node).
}
\description{
This function takes the average (Option="Mean") or the sum (Option="Sum") of all (observed) cells in a one-mode or two-mode network.
\itemize{
\item For a binary (0/1) network taking the sum gives the total number of ties present (value 1) in the network.
\item For a binary (0/1) network the average is equal to the density, i.e. the proportion of ties with values 1.
\item For a valued network (a network with weights) taking the sum gives the total weights of all cells in the matrix.
\item For a valued network taking the average means taking the total of all values in the matrix and divide this by the number of possible ties, which gives the average value for the observed cells in a network. Note that this includes values 0.
\item Cells with missing data (indicated by values 'NA' in the matrix) are ignored in both the numerator and denominator. Extra information about missing data is provided when running the function.
\item For one-mode networks you can also indicate whether to ignore the Loops entries (self-nominations). The default is to ignore the Loops (Loops=FALSE).
\item If an attribute file is specified for rows (ROWS) and/or the columns (COLS), then the average or sum is provided for each unique group in that attribute file. This might be useful to examine whether specific types of actors tend to have specific ties with specific other nodes. Note that ROWS and COLS does not have to contain the same attribute.
\item If ROWS or COLS is set to "Unique" (default is "All"), then the matrix is not aggregated for that row or column, and so values are obtained for each node. This might be useful if you want to know, for example, if specific nodes have more connections with specific types of actors.
}
}
\examples{

## A one-mode network
xDensity(Krackhardt_HighTech$Advice)
xDensity(Krackhardt_HighTech$Advice, ROWS=Krackhardt_HighTech$Attributes$Department,
   COLS=Krackhardt_HighTech$Attributes$Department)
xDensity(Krackhardt_HighTech$Advice, ROWS="All",
   COLS=Krackhardt_HighTech$Attributes$Department, Option="Sum")

## A two-mode network
xDensity(Davis_SouthernWomen$Attendance)
}
\references{
Chapter 10. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022).
Analyzing Social Networks Using R. SAGE.
}
\seealso{
\code{\link[=xDegreeCentrality]{xDegreeCentrality()}}, \code{\link[=xConnectedness]{xConnectedness()}}, \code{\link[=xCompactness]{xCompactness()}}, \code{\link[=xReciprocity]{xReciprocity()}}
}
