% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xDichotomize}
\alias{xDichotomize}
\title{Dichotomizes the ties of a network}
\usage{
xDichotomize(NET1, Value = 0, Type = "GT")
}
\arguments{
\item{NET1}{A network to be dichotomized.}

\item{Value}{The value at which to dichotomize. When using (Type="EQ") this can also be a vector of values to be set to 1.}

\item{Type}{The type of dichotomization. The five options are: "GT" (greater than), "LT" (less than), "EQ" (equal),}
}
\value{
A dichotomized network.
}
\description{
This function offers different options for dichotomizing a network. The default is "GT" with Value=0.
}
\examples{
M1<-matrix(c(1,2,3,4,5,6,7,8,NA),3,3)
xDichotomize(M1,Value=c(4,7,2,20),Type="EQ")
xDichotomize(M1,Value=c(4),Type="GT")
}
