% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{xdisplay}
\alias{xdisplay}
\title{Display a dataset in UCINET's Data|Display style}
\usage{
xdisplay(net, relation = NULL)
}
\arguments{
\item{net}{A network (any accepted form).}

\item{relation}{Which relation to show for a multi-relation dataset.}
}
\value{
The network, invisibly.
}
\description{
Display a dataset in UCINET's Data|Display style
}
