% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubgroupsCommunity.R
\name{xLabelPropagation}
\alias{xLabelPropagation}
\title{Provides the membership of nodes for the community detection using label propagation.}
\usage{
xLabelPropagation(
  NET1,
  vertex.shape = NULL,
  vertex.label.cex = NULL,
  layout = NULL
)
}
\arguments{
\item{NET1}{An undirected one-mode network stored as a 'matrix' object.}

\item{vertex.shape}{An option argument to specify the shapes of the nodes. Needs to be a vector of length equal to the number of nodes.}

\item{vertex.label.cex}{The size of the font used for the labels (default when NULL is cex=1).}

\item{layout}{The layout option used. See igraph. (default when NULL is layout_nicely, which tries to find the optimal layout).}
}
\value{
A vector with membership for each node (row).
}
\description{
Using an undirected one-mode network (NET1), this function calculates the solution for community detection membership using label propagation and graphically shows the result.
The output is a matrix providing membership for communities with (k) groups.
This function relies on igraph's cluster_label_prop() function.
}
\examples{
## Examples of undirected networks (See chapter 11 of Borgatti et al., 2022):
xLabelPropagation(Zachary_KarateClub$Connection, vertex.label.cex=.7)
## Add attribute file related to the club split:
## Define the shapes to be used
UseShapes<-c("circle","rectangle")
## Define the vector for nodes -
ZKAttrClub<-(Zachary_KarateClub$Attributes$Club>1)+1
xLabelPropagation(Zachary_KarateClub$Connection, vertex.shape=UseShapes[ZKAttrClub],
         vertex.label.cex=.7)

## A second dataset:
xLabelPropagation(Kapferer_TailorShop$SociationalT1, vertex.label.cex=.7,
                  layout="layout_with_fr")
}
\references{
Chapter 11. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
igraph package - https://igraph.org
}
\seealso{
\code{\link[=xGirvanNewman]{xGirvanNewman()}}, \code{\link[=xFastGreedy]{xFastGreedy()}}, \code{\link[=xWalkTrap]{xWalkTrap()}}, \code{\link[=xLouvainMethod]{xLouvainMethod()}}, \code{\link[igraph:cluster_label_prop]{igraph::cluster_label_prop()}}
}
