% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultivariateMethods.R
\name{xMDS}
\alias{xMDS}
\title{Performs multidimensional scaling on a matrix}
\usage{
xMDS(MAT1, Margins = 0.05, Distance = c(0, 4))
}
\arguments{
\item{MAT1}{A distance matrix}

\item{Margins}{The margins added to the output figure}

\item{Distance}{The distance of the names (labels) of the nodes to the midpoint (as a \% of the overall distance).}
}
\value{
A figure containing the MDS solution and the output from cmdscale.
}
\description{
Takes a distance matrix and performs multidimensional scaling on the matrix. This function relies on cmdscale function.
}
\examples{
CITIES1<-matrix(c(
     0,  206,  429, 1504,  963, 2976,
   206,    0,  233, 1308,  802, 2815,
   429,  233,    0, 1075,  671, 2684,
  1504, 1308, 1075,    0, 1329, 3273,
   963,  802,  671, 1329,    0, 2013,
  2976, 2815, 2684, 3273, 2013,    0),6,6)
colnames(CITIES1)<-c("BOSTON","NY","DC","MIAMI","CHICAGO","SEATTLE")
xMDS(CITIES1,Distance=c(0,4))
}
\references{
Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022).
Analyzing Social Networks Using R. SAGE.
}
