% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xNormalize}
\alias{xNormalize}
\title{Normalizes the values in a network}
\usage{
xNormalize(NET1, Type)
}
\arguments{
\item{NET1}{A network to be normalized.}

\item{Type}{The type of symmetrization. The five options are:
"Max" (divides the value of each cell value by the maximum of the tie values in a row),
"Av" (divides the value of each cell value by the average of the tie values in a row),
"Sum" (divides the value of each cell by the sum of the tie values in a row),
"Center" (subtracts the mean for a row from each cell value).}
}
\value{
The normalized network.
}
\description{
This function allows different forms of normalizing a network.
}
\examples{
M1<-matrix(c(1,2,3,4,5,6,7,8,9),3,3)
xNormalize(M1, Type="Max")
}
