% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{xRegression}
\alias{xRegression}
\title{Performs a regression analysis using a classic (and a permutation-based) significance test on a vector variable}
\usage{
xRegression(VEC1, DAT2, NPERM = 1000)
}
\arguments{
\item{VEC1}{A vector of values for the dependent variable.}

\item{DAT2}{A data.frame containing the independent variables.}

\item{NPERM}{A positive integer value indicating the number of permutations to be used for the statistical test. If NPERM is set to NULL, NA or 0, no permutation-based significance test is performed.}
}
\value{
Returns an object containing information about the model and the regression coefficients.
}
\description{
This function takes a vectors as a dependent variable and regresses it on a set of independent variables.
It returns the regression coefficients and the classic (two-tailed) significance tests.
If NPERM (number of permutations) is set to a value higher than 0 (default is 1000), permutations-based significance is also provided,
which will provide information about the regression coefficients for the permuted variables. A higher value than 1000 is generally advisable. Both the one-tailed ("AsSmall" or "AsLarge")
and the two-tailed ("AsExtreme") test are given. Note that cases with missing values are included in the permutation and this can impact the degrees of freedom for the regression analysis obtained from the permutations.
For example, suppose the same 20\% of cases are missing for both variables (i.e., unit non-response), then permuting these missings will mean that a lot more cases will have a missing for at least one of the variables.
\itemize{
\item Last updated: 23 April 2022.
}
}
\examples{
VEC1<-c(2,4,2,3,3,1)
X1<-c(4,2,3,2,1,5)
X2<-c(1,0,0,1,1,0)
DAT2<-data.frame(X1,X2)
RESULTS1<-xRegression(VEC1,DAT2,NPERM=1000)
RESULTS1
xRegression(VEC1,DAT2,NPERM=0)
}
\author{
Filip Agneessens \href{mailto:filipagneessens2@gmail.com}{filipagneessens2@gmail.com}
}
