% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EquivalenceRoles.R
\name{xStructuralEquivalence}
\alias{xStructuralEquivalence}
\title{Calculate structural equivalence for one or more networks}
\usage{
xStructuralEquivalence(
  NET1,
  Method = "Euclidean",
  IncludeTransposed = F,
  Choiceij = "OnlyReciprocal",
  DigitRound = 5
)
}
\arguments{
\item{NET1}{A matrix or list of matrices, which could be part of a xUCINET project}

\item{Method}{Method used to obtain the similarity/distances between two nodes, i and j, which can be based on:
\itemize{
\item "AbsDiff" the sum of the absolute differences between both nodes in their relations to others. I.e.,
the value for the pair of nodes i,j is given by: |(i->k)-(j->k)| taken over all k. In the case where also
the incoming ties are considered (see IncludeTransposed) this also includes: |(k->i)-(k->j)| taken over
all k. The output is a distance matrix.
\item "Euclidean" the squared square root of the sum of the squares of the differences between
corresponding values. I.e., the sum of the value for the pair of nodes i,j is given by: ((i->k)-(j->k))^2
taken over all k (as well as ((k->i)-(k->j))^2 in the case when also incoming ties are considered).
For the resulting sum for each pair i,j the square root is then taken. The output is a distance matrix.
\item "MatchesN" the number of exact matches, i.e., the number of times that i and j relate to other nodes
k with exactly the same value. (i->k)==(j->k) taken over all k (as well as (k->i)==(k->j) if the incoming
ties are also considered). The output is a similarity matrix.
\item "PosMatchesN" the number of exact positive matches, i.e., the number of times that i and j relate to
other nodes k with exactly the same positive value (i.e., value different from 0). This approach differs
from "MatchesN" in that it excludes similarity based on two nodes not being connected to others. The output
is a similarity matrix. For a nimary network it provides an index of the number of common non-zero choices.
\item  "Product" the product of the link between i and k, and j and k. I.e., (i->k)*(j->k) taken over all k
(and (k->i)-(k->j) if the incoming ties are also considered). The output is a similarity matrix with high
values indicating both i and j are highly connected to k.
\item  "Pearson" takes the values for the links between i and k, and correlates it with the values for the
corresponding links between j and k. I.e., cor ((i->k),(j->k)) for outgoing ties.  The output is a similarity
matrix.
\item  "Kendall" similar to "Pearson", except it uses Kendall's Tau.
\item  "Spearman" similar to "Pearson", except it uses Spearman's Rho, i.e., the ranked value.
}}

\item{IncludeTransposed}{Whether or not to focus only on outgoing ties to define the structural equivalence
between i and j (based on the measure defined in Method. If FALSE then only the outgoing ties to other
nodes (k) are considered (i.e., i->k and j->k) when calculating the structural equivalence between nodes
i and j, while if TRUE the also the incoming ties to i and j from other nodes (i.e., k->i and k->j) will
be used to calculate the structural equivalence between nodes i and j.}

\item{Choiceij}{What to do with ties between i and j and selfloops?
\itemize{
\item If "Ignore", then only the ties from i and j with other nodes k are considered (where k is
not i or j), i.e., the values for i->i, j->i, i->j and j->j are ignored.
\item If "OnlyReciprocal", then loops (i->i and j->j) are ignored, and i->j is compared to j->i.
\item If "Reciprocal", then i->j is compared to j->i and i->i with j->j.
\item If "Original", then the classic approach of comparing rows/columns is used, and hence i->i is
compared to j->i, while i->j is compared to j->j (plus i->i with i->j and j->i with j->j when incoming
ties are also considering)
}}

\item{DigitRound}{Whether to round the output to a specified number of decimal places (default = 5).}
}
\value{
A matrix containing the level of similarity between each pair of nodes (or distances in the case of
choosing the euclidean distance or absolute difference).
}
\description{
Computes for each pair of nodes (i and j) the level of structural equivalence based upon comparisons of rows
(and potentially also columns) in the adjacency matrix (or whatever else is used as input matrix).
Structural equivalence is based on a pair of nodes (i and j) being in the same/in a similar way connected to
other nodes (k).
Multiple relations are permissible as input (which should be combined in a list) and in such a case the
combined similarity/distance across these different relations are considered. The matrix can contain values or
binary data.
When consider a pair of nodes (i and j), the measure can include or exclude reciprocal connections (i->j and j->i)
as well as selfloops (i->i) and (j->j).
Calculations of similarity can be based on a a variety of measures, including Pearson correlation, exact matches
or matches of positive matches only.
Note that measures based on absolute difference and Euclidean distances produces a distance matrix, while all
the other options produce a similarity matrix.
}
\examples{
## A small example with 3 networks focusing on outgoing ties only (default):
xStructuralEquivalence(ASNR_Fig12x1R1)
xStructuralEquivalence(list(ASNR_Fig12x1R1,ASNR_Fig12x1R2,ASNR_Fig12x1R3))

## Incoming and outgoing ties:
xStructuralEquivalence(list(ASNR_Fig12x1R1,ASNR_Fig12x1R2,ASNR_Fig12x1R3))

## Apply the function using both incoming and outgoing ties for esteem and disesteem
## (Sampson Monastery):
Samp1<-list(Sampson_Monastery$Esteem, Sampson_Monastery$Disesteem)
xStructuralEquivalence(Samp1,IncludeTransposed=TRUE)
}
\references{
Chapter 12. Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022).
Analyzing Social Networks Using R. SAGE.
}
