% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeLevelMeasures.R
\name{xStructuralHoles}
\alias{xStructuralHoles}
\title{Measures to capture structural holes.}
\usage{
xStructuralHoles(
  NET1,
  Measures = c("EffectiveSize", "Constraint"),
  Loops = FALSE,
  Include = "WholeNetwork",
  Direction = "Out"
)
}
\arguments{
\item{NET1}{A one-mode network dataset stored as an object of class "matrix".}

\item{Measures}{The choice of output to be provided. This can be "EffectiveSize", and/or "Constraint".}

\item{Loops}{Whether to include the Loops (default is to ignore the Loops).}

\item{Include}{Whether to consider the whole network ("WholeNetwork") or only the egonetwork ("EgoNetwork") around the node.}

\item{Direction}{How to define the alters for a node (at the moment only outgoing choices).}
}
\value{
A matrix containing different measures (columns) for each node (row).
}
\description{
This function takes a one-mode network (\code{NET1}) and calculates measures of structural equivalence.
It returns for each node in the rows of the matrix information about effective size and/or constraint.
Both measures also work for valued networks.
}
\examples{
## Consider the Friendship network for Krackhardt's high-tech managers.
xStructuralHoles(Padgett_FlorentineFamilies$Marriage)
## Consider the Friendship network for Krackhardt's high-tech managers.
xStructuralHoles(Krackhardt_HighTech$Friendship)
}
