% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xSymmetrize}
\alias{xSymmetrize}
\title{Symmetrizes the ties of a network}
\usage{
xSymmetrize(NET1, Type = "Min")
}
\arguments{
\item{NET1}{A network to be symmetrized.}

\item{Type}{The type of symmetrization. The five options are: "Min" (take the minimum of both ties), "Max" (take the maximum of both ties),
"Av" (take the average of both ties), "Sum" (take the sum of both ties), "Prod" (take the product of both ties).}
}
\value{
The symmetrized network.
}
\description{
This function allows different forms of symmetrization on a network.
}
\examples{
M1<-matrix(c(1,2,3,4,5,6,7,8,NA),3,3)
xSymmetrize(M1)
}
