% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{xTranspose}
\alias{xTranspose}
\title{Transposes the ties of a network}
\usage{
xTranspose(NET1)
}
\arguments{
\item{NET1}{A network to be transposed.}
}
\value{
The transposed network.
}
\description{
This function transposes a network containing network data.
}
\examples{
M1<-matrix(c(1,2,3,4,5,6,7,8,NA),3,3)
xTranspose(M1)
}
