% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NodeLevelMeasures.R
\name{xValuedTieComposition}
\alias{xValuedTieComposition}
\title{Provides different measures about the outgoing ties for nodes on a valued network}
\usage{
xValuedTieComposition(
  NET1,
  Measures = c("SumStrength", "AvStrength", "SDStrength", "MedianStrength",
    "RangeStrength", "MinStrength", "MaxStrength", "IQRStrength", "Q1Strength",
    "Q3Strength", "IQVStrength", "DistrStrength"),
  Loops = FALSE,
  Mode = "OneMode",
  Exclude0 = FALSE,
  NrCategoriesIQV = 1
)
}
\arguments{
\item{NET1}{A matrix stored as an object of class matrix}

\item{Measures}{The measures to be calculated about the set of networks. Options are: "SumStrength","AvStrength","SDStrength","MedianStrength","RangeStrength","MinStrength","MaxStrength","IQRStrength","Q1Strength","Q3Strength","IQVStrength", and "DistrStrength".}

\item{Loops}{Logical, whether the diagonal values should be ignored (default setting is FALSE, i.e., to ignore the diagonal values)}

\item{Mode}{Whether the networks is one-mode "OneMode" or two-mode networks "TwoMode".}

\item{Exclude0}{Whether to exclude values 0 from the analysis (default is FALSE).}

\item{NrCategoriesIQV}{An option value to be used for IQV strength, in case there are some additional values/categories that should be considered, but might not be present in the data. A value 1 means that the observed number of categories should be used.}
}
\value{
A matrix with different measures (columns) for each node (row)
}
\description{
Takes an object of class 'matrix', which is a valued network, and calculates different measures:
\itemize{
\item the sum of the number of valid alters, i.e., the number of alters for which information is available (not equal to NA),
\item the sum of tie strengths to alters (Measures="SumStrength"),
\item the average tie strengths to alters relative to the number of valid cases (Measures="AvStrength"),
\item the standard deviation of tie strengths to alters (Measures="SDStrength"),
\item the median of tie strengths to alters (Measures="MedianStrength"),
\item the range of tie strengths to alters, i.e., the maximum minus the minimum value (Measures="RangeStrength"),
\item the minimum tie strengths to alters (Measures="MinStrength"),
\item the maximum tie strengths to alters (Measures="MaxStrength"),
\item the interquartile for tie strengths to alters (Measures="IQRStrength"),
\item the first quartile for tie strengths to alters (Measures="Q1Strength"),
\item the third quartile for tie strengths to alters (Measures="Q3Strength"),
\item the IQV for tie strengths to alters (Measures="IQVStrength"),
\item the distribution of tie strengths to alters (Measures="DistrStrength").
}
}
\examples{
## Applied to the valued network for Zachary's Karate club:
xValuedTieComposition(Zachary_KarateClub$Strength, Exclude0=TRUE, NrCategoriesIQV = 20)
## See Chapter 8 in Borgatti, Everett, Johnson and Agneessens (2022)
}
\references{
Chapter 8 in Borgatti, S. P., Everett, M. G., Johnson, J. C., & Agneessens, F. (2022). Analyzing Social Networks Using R. SAGE.
}
